@@ -22,7 +22,6 @@ from typing import (
     Union, cast,
 )
 
-import psycopg2.extensions
 import werkzeug.exceptions
 from werkzeug import Response
 
@@ -34,19 +33,23 @@ from cdedb.common import (
     CdEDBObject, CdEDBObjectMap, CdEDBOptionalMap, CourseChoiceToolActions,
     CourseFilterPositions, DefaultReturnCode, EntitySorter, Error, InfiniteEnum,
     KeyFunction, LodgementsSortkeys, PartialImportError, RequestState, Sortkey,
-    asciificator, deduct_years, determine_age_class, diacritic_patterns, get_hash, glue,
-    json_serialize, merge_dicts, mixed_existence_sorter, n_, now, unwrap, xsorted,
+    asciificator, deduct_years, determine_age_class, diacritic_patterns, get_hash,
+    get_localized_country_codes, json_serialize, merge_dicts,
+    mixed_existence_sorter, n_, now, unwrap, xsorted,
 )
-from cdedb.database.connection import Atomizer
 from cdedb.filter import (
-    date_filter, enum_entries_filter, keydictsort_filter, money_filter, safe_filter,
+    enum_entries_filter, keydictsort_filter, safe_filter,
 )
 from cdedb.frontend.common import (
     AbstractUserFrontend, CustomCSVDialect, RequestConstraint, REQUESTdata,
     REQUESTdatadict, REQUESTfile, access, calculate_db_logparams, calculate_loglinks,
     cdedbid_filter, cdedburl, check_validation as check,
     check_validation_optional as check_optional, event_guard, make_event_fee_reference,
-    periodic, process_dynamic_input, request_extractor,
+    periodic, process_dynamic_input, request_extractor, make_persona_name,
+    TransactionObserver,
+)
+from cdedb.frontend.event_lodgement_wishes import (
+    create_lodgement_wishes_graph, detect_lodgement_wishes,
 )
 from cdedb.query import (
     Query, QueryConstraint, QueryOperators, QueryScope, make_registration_query_aux,
@@ -55,6 +58,7 @@ from cdedb.query import (
 from cdedb.validation import (
     COURSE_COMMON_FIELDS, EVENT_EXPOSED_FIELDS, LODGEMENT_COMMON_FIELDS,
     PERSONA_FULL_EVENT_CREATION, TypeMapping, filter_none, validate_check,
+    EVENT_PART_COMMON_FIELDS, EVENT_PART_CREATION_MANDATORY_FIELDS
 )
 from cdedb.validationtypes import VALIDATOR_LOOKUP
 
@@ -169,7 +173,7 @@ class EventFrontend(AbstractUserFrontend):
                 enum_entries_filter(
                     const.Genders,
                     rs.gettext if download is None else rs.default_gettext)),
-            'country': OrderedDict(self.get_localized_country_codes(rs)),
+            'country': OrderedDict(get_localized_country_codes(rs)),
         }
         return self.generic_user_search(
             rs, download, is_search, QueryScope.event_user, QueryScope.event_user,
@@ -309,7 +313,7 @@ class EventFrontend(AbstractUserFrontend):
         as well."""
         course_ids = self.eventproxy.list_courses(rs, event_id)
         courses = self.eventproxy.get_courses(rs, course_ids)
-        registration_ids = self.eventproxy.list_registrations(rs, event_id)
+        registration_ids = self.eventproxy.list_participants(rs, event_id)
         registrations = self.eventproxy.get_registrations(rs, registration_ids)
 
         if not part_ids:
@@ -544,43 +548,47 @@ class EventFrontend(AbstractUserFrontend):
                       {'address': ml_address})
         return self.redirect(rs, "event/show_event")
 
-    @access("event")
-    @event_guard()
-    def part_summary_form(self, rs: RequestState, event_id: int) -> Response:
-        """Render form."""
-        tracks = rs.ambience['event']['tracks']
-        current = {
-            f"{key}_{part_id}": value
-            for part_id, part in rs.ambience['event']['parts'].items()
-            for key, value in part.items() if key not in ('id', 'tracks')}
-        for part_id, part in rs.ambience['event']['parts'].items():
-            for track_id, track in part['tracks'].items():
-                for k in ('title', 'shortname', 'num_choices', 'min_choices',
-                          'sortkey'):
-                    current[f"track_{k}_{part_id}_{track_id}"] = track[k]
-            for m_id, m in part['fee_modifiers'].items():
-                for k in ('modifier_name', 'amount', 'field_id'):
-                    current[f"fee_modifier_{k}_{part_id}_{m_id}"] = m[k]
-        merge_dicts(rs.values, current)
-        referenced_parts: Set[int] = set()
-        referenced_tracks: Set[int] = set()
-        has_registrations = self.eventproxy.has_registrations(rs, event_id)
+    def _deletion_blocked_parts(self, rs: RequestState, event_id: int) -> Set[int]:
+        """Returns all part_ids from parts of a given event which must not be deleted.
+
+        Extracts all parts of the given event from the database and checks if there are
+        blockers preventing their deletion.
+
+        :returns: All part_ids whose deletion is blocked.
+        """
+        blocked_parts: Set[int] = set()
+        if len(rs.ambience['event']['parts']) == 1:
+            blocked_parts.add(unwrap(rs.ambience['event']['parts'].keys()))
+        course_ids = self.eventproxy.list_courses(rs, event_id)
+        courses = self.eventproxy.get_courses(rs, course_ids.keys())
+        # referenced tracks block part deletion
+        for course in courses.values():
+            for track_id in course['segments']:
+                blocked_parts.add(rs.ambience['event']['tracks'][track_id]['part_id'])
+        return blocked_parts
+
+    def _deletion_blocked_tracks(self, rs: RequestState, event_id: int) -> Set[int]:
+        """Returns all track_ids from tracks of a given event which must not be deleted.
+
+        Extracts all tracks of the given event from the database and checks if there are
+        blockers preventing their deletion.
+
+        :returns: All track_ids whose deletion is blocked.
+        """
+        blocked_tracks: Set[int] = set()
         course_ids = self.eventproxy.list_courses(rs, event_id)
         courses = self.eventproxy.get_courses(rs, course_ids.keys())
         for course in courses.values():
-            referenced_tracks.update(course['segments'])
-        # referenced tracks block part deletion
-        for track_id in referenced_tracks:
-            referenced_parts.add(tracks[track_id]['part_id'])
+            blocked_tracks.update(course['segments'])
+        return blocked_tracks
 
-        sorted_fields = xsorted(rs.ambience['event']['fields'].values(),
-                                key=EntitySorter.event_field)
-        legal_datatypes, legal_assocs = EVENT_FIELD_SPEC['fee_modifier']
-        fee_modifier_fields = [
-            (field['id'], field['field_name']) for field in sorted_fields
-            if field['association'] in legal_assocs
-            and field['kind'] in legal_datatypes
-        ]
+    @access("event")
+    @event_guard()
+    def part_summary(self, rs: RequestState, event_id: int) -> Response:
+        """Display a comprehensive overview of all parts of a given event."""
+        has_registrations = self.eventproxy.has_registrations(rs, event_id)
+        referenced_parts = self._deletion_blocked_parts(rs, event_id)
+
         fee_modifiers_by_part = {
             part_id: {
                 e['id']: e
@@ -589,265 +597,195 @@ class EventFrontend(AbstractUserFrontend):
             }
             for part_id in rs.ambience['event']['parts']
         }
+
+        return self.render(rs, "part_summary", {
+            'fee_modifiers_by_part': fee_modifiers_by_part,
+            'referenced_parts': referenced_parts,
+            'has_registrations': has_registrations})
+
+    @access("event", modi={"POST"})
+    @event_guard()
+    @REQUESTdata("ack_delete")
+    def delete_part(self, rs: RequestState, event_id: int, part_id: int,
+                    ack_delete: bool) -> Response:
+        """Delete a given part."""
+        if not ack_delete:
+            rs.append_validation_error(
+                ("ack_delete", ValueError(n_("Must be checked."))))
+        if rs.has_validation_errors():
+            return self.part_summary(rs, event_id)
+        if self.eventproxy.has_registrations(rs, event_id):
+            raise ValueError(n_("Registrations exist, no deletion."))
+        if part_id in self._deletion_blocked_parts(rs, event_id):
+            raise ValueError(n_("This part can not be deleted."))
+
+        event = {
+            'id': event_id,
+            'parts': {part_id: None},
+        }
+        code = self.eventproxy.set_event(rs, event)
+        self.notify_return_code(rs, code)
+
+        return self.redirect(rs, "event/part_summary")
+
+    @access("event")
+    @event_guard()
+    def add_part_form(self, rs: RequestState, event_id: int) -> Response:
+        if self.eventproxy.has_registrations(rs, event_id):
+            rs.notify("error", n_("Registrations exist, no part creation possible."))
+            return self.redirect(rs, "event/show_event")
+        sorted_fields = xsorted(rs.ambience['event']['fields'].values(),
+                                key=EntitySorter.event_field)
         legal_datatypes, legal_assocs = EVENT_FIELD_SPEC['waitlist']
         waitlist_fields = [
             (field['id'], field['field_name']) for field in sorted_fields
-            if field['association'] in legal_assocs
-            and field['kind'] in legal_datatypes
+            if field['association'] in legal_assocs and field['kind'] in legal_datatypes
         ]
-        return self.render(rs, "part_summary", {
-            'fee_modifier_fields': fee_modifier_fields,
-            'fee_modifiers_by_part': fee_modifiers_by_part,
+        return self.render(rs, "add_part", {
             'waitlist_fields': waitlist_fields,
-            'referenced_parts': referenced_parts,
+            'DEFAULT_NUM_COURSE_CHOICES': DEFAULT_NUM_COURSE_CHOICES})
+
+    @access("event", modi={"POST"})
+    @event_guard()
+    @REQUESTdatadict(*EVENT_PART_CREATION_MANDATORY_FIELDS)
+    def add_part(self, rs: RequestState, event_id: int, data: CdEDBObject) -> Response:
+        if self.eventproxy.has_registrations(rs, event_id):
+            raise ValueError(n_("Registrations exist, no part creation possible."))
+
+        data = check(rs, vtypes.EventPart, data)
+        if rs.has_validation_errors():
+            return self.add_part_form(rs, event_id)
+        assert data is not None
+
+        # check non-static dependencies
+        if data["waitlist_field"]:
+            waitlist_field = rs.ambience['event']['fields'][data["waitlist_field"]]
+            allowed_datatypes, allowed_associations = EVENT_FIELD_SPEC['waitlist']
+            if (waitlist_field['association'] not in allowed_associations
+                    or waitlist_field['kind'] not in allowed_datatypes):
+                rs.append_validation_error(("waitlist_field", ValueError(
+                    n_("Waitlist linked to non-fitting field."))))
+        if rs.has_validation_errors():
+            return self.add_part_form(rs, event_id)
+
+        event = {'id': event_id, 'parts': {-1: data}}
+        code = self.eventproxy.set_event(rs, event)
+        self.notify_return_code(rs, code)
+
+        return self.redirect(rs, "event/part_summary")
+
+    @access("event")
+    @event_guard()
+    def change_part_form(self, rs: RequestState, event_id: int, part_id: int
+                         ) -> Response:
+        part = rs.ambience['event']['parts'][part_id]
+
+        current = copy.deepcopy(part)
+        del current['id']
+        del current['tracks']
+        for track_id, track in part['tracks'].items():
+            for k in ('title', 'shortname', 'num_choices', 'min_choices', 'sortkey'):
+                current[f"track_{k}_{track_id}"] = track[k]
+        for m in rs.ambience['event']['fee_modifiers'].values():
+            for k in ('modifier_name', 'amount', 'field_id'):
+                current[f"fee_modifier_{k}_{m['id']}"] = m[k]
+        merge_dicts(rs.values, current)
+
+        has_registrations = self.eventproxy.has_registrations(rs, event_id)
+        referenced_tracks = self._deletion_blocked_tracks(rs, event_id)
+
+        sorted_fields = xsorted(rs.ambience['event']['fields'].values(),
+                                key=EntitySorter.event_field)
+        legal_datatypes, legal_assocs = EVENT_FIELD_SPEC['fee_modifier']
+        fee_modifier_fields = [
+            (field['id'], field['field_name']) for field in sorted_fields
+            if field['association'] in legal_assocs and field['kind'] in legal_datatypes
+        ]
+        fee_modifiers = {
+            e['id']: e
+            for e in rs.ambience['event']['fee_modifiers'].values()
+            if e['part_id'] == part_id
+        }
+        legal_datatypes, legal_assocs = EVENT_FIELD_SPEC['waitlist']
+        waitlist_fields = [
+            (field['id'], field['field_name']) for field in sorted_fields
+            if field['association'] in legal_assocs and field['kind'] in legal_datatypes
+        ]
+        return self.render(rs, "change_part", {
+            'part_id': part_id,
+            'fee_modifier_fields': fee_modifier_fields,
+            'fee_modifiers': fee_modifiers,
+            'waitlist_fields': waitlist_fields,
             'referenced_tracks': referenced_tracks,
             'has_registrations': has_registrations,
             'DEFAULT_NUM_COURSE_CHOICES': DEFAULT_NUM_COURSE_CHOICES})
 
-    @staticmethod
-    def process_part_input(rs: RequestState, has_registrations: bool
-                           ) -> Dict[int, Optional[CdEDBObject]]:
-        """This handles input to configure the parts.
+    @access("event", modi={"POST"})
+    @event_guard(check_offline=True)
+    @REQUESTdatadict(*EVENT_PART_COMMON_FIELDS)
+    def change_part(self, rs: RequestState, event_id: int, part_id: int,
+                    data: CdEDBObject) -> Response:
+        """Change one part, including the associated tracks and fee modifiers."""
+        # this will be added at the end after processing the dynamic input and will only
+        # yield false validation errors
+        del data['tracks']
+        del data['fee_modifiers']
+        data = check(rs, vtypes.EventPart, data)
+        if rs.has_validation_errors():
+            return self.change_part_form(rs, event_id, part_id)
+        assert data is not None
+        has_registrations = self.eventproxy.has_registrations(rs, event_id)
 
-        Since this covers a variable number of rows, we cannot do this
-        statically. This takes care of validation too.
-        """
-        parts = rs.ambience['event']['parts']
-        fee_modifiers = rs.ambience['event']['fee_modifiers']
+        #
+        # Check part specific stuff which can not be checked statically
+        #
+        if data["waitlist_field"]:
+            waitlist_field = rs.ambience['event']['fields'][data["waitlist_field"]]
+            allowed_datatypes, allowed_associations = EVENT_FIELD_SPEC['waitlist']
+            if (waitlist_field['association'] not in allowed_associations
+                    or waitlist_field['kind'] not in allowed_datatypes):
+                rs.append_validation_error(("waitlist_field", ValueError(
+                    n_("Waitlist linked to non-fitting field."))))
 
-        # Handle basic part data
-        delete_flags = request_extractor(
-            rs, {f"delete_{part_id}": bool for part_id in parts})
-        deletes = {part_id for part_id in parts
-                   if delete_flags['delete_{}'.format(part_id)]}
-        if has_registrations and deletes:
-            raise ValueError(n_("Registrations exist, no deletion."))
-        spec: TypeMapping = {
+        #
+        # process the dynamic track input
+        #
+        def track_constraint_maker(track_id: int, prefix: str
+                                   ) -> List[RequestConstraint]:
+            min_choice = f"{prefix}min_choices_{track_id}"
+            num_choice = f"{prefix}num_choices_{track_id}"
+            msg = n_("Must be less or equal than total Course Choices.")
+            return [(
+                lambda d: d[min_choice] <= d[num_choice], (min_choice, ValueError(msg))
+            )]
+
+        track_existing = rs.ambience['event']['parts'][part_id]['tracks']
+        track_spec = {
             'title': str,
             'shortname': str,
-            'part_begin': datetime.date,
-            'part_end': datetime.date,
-            'fee': decimal.Decimal,
-            'waitlist_field': Optional[vtypes.ID],  # type: ignore
+            'num_choices': vtypes.NonNegativeInt,
+            'min_choices': vtypes.NonNegativeInt,
+            'sortkey': int
         }
-        params: TypeMapping = {
-            f"{key}_{part_id}": value
-            for part_id in parts if part_id not in deletes
-            for key, value in spec.items()
-        }
+        track_data = process_dynamic_input(
+            rs, track_existing, track_spec, prefix="track_",
+            constraint_maker=track_constraint_maker)
 
-        # noinspection PyRedundantParentheses
-        def part_constraint_maker(part_id: int) -> List[RequestConstraint]:
-            begin = f"part_begin_{part_id}"
-            end = f"part_end_{part_id}"
-            msg = n_("Must be later than begin.")
-            ret: List[RequestConstraint]
-            ret = [(lambda d: d[begin] <= d[end], (end, ValueError(msg)))]
+        deleted_tracks = {anid for anid in track_data if track_data[anid] is None}
+        new_tracks = {anid for anid in track_data if anid < 0}
+        if deleted_tracks and has_registrations:
+            raise ValueError(n_("Registrations exist, no track deletion possible."))
+        if deleted_tracks & self._deletion_blocked_tracks(rs, event_id):
+            raise ValueError(n_("Some tracks can not be deleted."))
+        if new_tracks and has_registrations:
+            raise ValueError(n_("Registrations exist, no track creation possible."))
 
-            key = f"waitlist_field_{part_id}"
-            fields = rs.ambience['event']['fields']
-            legal_datatypes, legal_assocs = EVENT_FIELD_SPEC['waitlist']
-            ret.append((
-                lambda d: d[key] is None or
-                          fields[d[key]]['association'] in legal_assocs
-                          and fields[d[key]]['kind'] in legal_datatypes,
-                (key, ValueError(n_(
-                    "Waitlist linked to non-fitting field.")))))
-            return ret
-
-        constraints = list(itertools.chain.from_iterable(
-            part_constraint_maker(part_id)
-            for part_id in parts if part_id not in deletes))
-        data = request_extractor(rs, params, constraints)
-        ret: Dict[int, CdEDBObject] = {
-            part_id: {key: data["{}_{}".format(key, part_id)] for key in spec}
-            for part_id in parts if part_id not in deletes
-        }
-
-        def track_params(part_id: int, track_id: int) -> TypeMapping:
-            """
-            Helper function to create the parameter extraction configuration
-            for the data of a single track.
-            """
-            return {
-                f"track_{k}_{part_id}_{track_id}": t
-                for k, t in {
-                    'title': str,
-                    'shortname': str,
-                    'num_choices': vtypes.NonNegativeInt,
-                    'min_choices': vtypes.NonNegativeInt,
-                    'sortkey': int
-                }.items()
-            }
-
-        def track_excavator(req_data: CdEDBObject, part_id: int, track_id: int
-                            ) -> CdEDBObject:
-            """
-            Helper function to create a single track's data dict from the
-            extracted request data.
-            """
-            return {
-                k: req_data['track_{}_{}_{}'.format(k, part_id, track_id)]
-                for k in ('title', 'shortname', 'num_choices', 'min_choices',
-                          'sortkey')}
-
-        # Handle newly created parts
-        marker = 1
-        while marker < 2 ** 10:
-            will_create = unwrap(request_extractor(
-                rs, {f"create_-{marker}": bool}))
-            if will_create:
-                if has_registrations:
-                    raise ValueError(n_("Registrations exist, no creation."))
-                params = {f"{key}_-{marker}": value for key, value in spec.items()}
-                constraints = part_constraint_maker(-marker)
-                data = request_extractor(rs, params, constraints)
-                ret[-marker] = {key: data[f"{key}_-{marker}"] for key in spec}
-            else:
-                break
-            marker += 1
-        # Return index of last new row to template to generate all inputs
-        # previously added by JS
-        rs.values['create_last_index'] = marker - 1
-
-        # Handle track data
-        track_delete_flags = request_extractor(rs, {
-            f"track_delete_{part_id}_{track_id}": bool
-            for part_id, part in parts.items()
-            for track_id in part['tracks']
-        })
-        track_deletes = {
-            track_id
-            for part_id, part in parts.items() for track_id in part['tracks']
-            if track_delete_flags['track_delete_{}_{}'.format(part_id,
-                                                              track_id)]
-        }
-        if has_registrations and track_deletes:
-            raise ValueError(n_("Registrations exist, no deletion."))
-        params = dict(itertools.chain.from_iterable(map(lambda d: d.items(),
-            (
-                track_params(part_id, track_id)
-                for part_id, part in parts.items()
-                for track_id in part['tracks']
-                if track_id not in track_deletes
-            )
-        )))
-
-        # noinspection PyRedundantParentheses
-        def track_constraint_maker(part_id: int, track_id: int) -> RequestConstraint:
-            # pylint: disable=redefined-builtin
-            min = "track_min_choices_{}_{}".format(part_id, track_id)
-            num = "track_num_choices_{}_{}".format(part_id, track_id)
-            msg = n_("Must be less or equal than total Course Choices.")
-            return (lambda d: d[min] <= d[num], (min, ValueError(msg)))
-
-        constraints = list(
-            track_constraint_maker(part_id, track_id)
-            for part_id, part in parts.items()
-            for track_id in part['tracks']
-            if track_id not in track_deletes)
-        data = request_extractor(rs, params, constraints)
-        rs.values['track_create_last_index'] = {}
-        for part_id, part in parts.items():
-            if part_id in deletes:
-                continue
-            ret[part_id]['tracks'] = {
-                track_id: (track_excavator(data, part_id, track_id)
-                           if track_id not in track_deletes else None)
-                for track_id in part['tracks']}
-            marker = 1
-            while marker < 2 ** 5:
-                will_create = unwrap(request_extractor(
-                    rs,
-                    {f"track_create_{part_id}_-{marker}": bool}))
-                if will_create:
-                    if has_registrations:
-                        raise ValueError(
-                            n_("Registrations exist, no creation."))
-                    params = track_params(part_id, -marker)
-                    constraints = [track_constraint_maker(part_id, -marker)]
-                    newtrack = track_excavator(
-                        request_extractor(rs, params, constraints),
-                        part_id, -marker)
-                    ret[part_id]['tracks'][-marker] = newtrack
-                else:
-                    break
-                marker += 1
-            rs.values['track_create_last_index'][part_id] = marker - 1
-
-        # And now track data for newly created parts
-        for new_part_id in range(1, rs.values['create_last_index'] + 1):
-            ret[-new_part_id]['tracks'] = {}
-            marker = 1
-            while marker < 2 ** 5:
-                will_create = unwrap(request_extractor(
-                    rs,
-                    {f"track_create_-{new_part_id}_-{marker}": bool}))
-                if will_create:
-                    params = track_params(-new_part_id, -marker)
-                    constraints = [
-                        track_constraint_maker(-new_part_id, -marker)]
-                    newtrack = track_excavator(
-                        request_extractor(rs, params, constraints),
-                        -new_part_id, -marker)
-                    ret[-new_part_id]['tracks'][-marker] = newtrack
-                else:
-                    break
-                marker += 1
-            rs.values['track_create_last_index'][-new_part_id] = marker - 1
-
-        def fee_modifier_params(part_id: int, fee_modifier_id: int) -> TypeMapping:
-            """
-            Helper function to create the parameter extraction configuration
-            for the data of a single fee modifier.
-            """
-            return {
-                f"fee_modifier_{k}_{part_id}_{fee_modifier_id}": t
-                for k, t in {
-                    'modifier_name': vtypes.RestrictiveIdentifier,
-                    'amount': decimal.Decimal,
-                    'field_id': vtypes.ID,
-                }.items()
-            }
-
-        def fee_modifier_excavator(req_data: CdEDBObject, part_id: int,
-                                   fee_modifier_id: int) -> CdEDBObject:
-            """
-            Helper function to create a single fee modifier's data dict from the
-            extracted request data.
-            """
-            ret = {
-                k: req_data['fee_modifier_{}_{}_{}'.format(
-                    k, part_id, fee_modifier_id)]
-                for k in ('modifier_name', 'amount', 'field_id')}
-            ret['part_id'] = part_id
-            if fee_modifier_id > 0:
-                ret['id'] = fee_modifier_id
-            return ret
-
-        # Handle fee modifier data
-        fee_modifier_delete_flags = request_extractor(
-            rs, {f"fee_modifier_delete_{mod['part_id']}_{mod['id']}": bool
-                 for mod in fee_modifiers.values()})
-        fee_modifier_deletes = {
-            mod['id']
-            for mod in fee_modifiers.values()
-            if fee_modifier_delete_flags['fee_modifier_delete_{}_{}'.format(
-                mod['part_id'], mod['id'])]
-        }
-        if has_registrations and fee_modifier_deletes:
-            raise ValueError(n_("Registrations exist, no deletion."))
-        params = dict(itertools.chain.from_iterable(map(lambda d: d.items(),
-            (
-                fee_modifier_params(mod['part_id'], mod['id'])
-                for mod in fee_modifiers.values()
-                if mod['id'] not in fee_modifier_deletes)
-            )
-        ))
-
-        def constraint_maker(part_id: int, fee_modifier_id: int
-                             ) -> List[RequestConstraint]:
-            key = f"fee_modifier_field_id_{part_id}_{fee_modifier_id}"
+        #
+        # process the dynamic fee modifier input
+        #
+        def fee_modifier_constraint_maker(
+                fee_modifier_id: int, prefix: str) -> List[RequestConstraint]:
+            key = f"{prefix}field_id_{fee_modifier_id}"
             fields = rs.ambience['event']['fields']
             legal_datatypes, legal_assocs = EVENT_FIELD_SPEC['fee_modifier']
             msg = n_("Fee Modifier linked to non-fitting field.")
@@ -857,122 +795,62 @@ class EventFrontend(AbstractUserFrontend):
                 (key, ValueError(msg))
             )]
 
-        constraints = list(itertools.chain.from_iterable(
-            constraint_maker(mod['part_id'], mod['id'])
-            for mod in fee_modifiers.values()
-            if mod['id'] not in fee_modifier_deletes))
+        fee_modifier_existing = [
+            mod['id'] for mod in rs.ambience['event']['fee_modifiers'].values()
+            if mod['part_id'] == part_id
+        ]
+        fee_modifier_spec = {
+            'modifier_name': vtypes.RestrictiveIdentifier,
+            'amount': decimal.Decimal,
+            'field_id': vtypes.ID,
+        }
+        fee_modifier_prefix = "fee_modifier_"
+        # do not change fee modifiers once registrations exist
+        if has_registrations:
+            fee_modifier_data = dict()
+        else:
+            fee_modifier_data = process_dynamic_input(
+                rs, fee_modifier_existing, fee_modifier_spec,
+                prefix=fee_modifier_prefix,
+                constraint_maker=fee_modifier_constraint_maker)
 
-        data = request_extractor(rs, params, constraints)
-        rs.values['fee_modifier_create_last_index'] = {}
-        ret_fee_modifiers: Dict[int, Optional[CdEDBObject]] = {
-            mod['id']: (fee_modifier_excavator(data, mod['part_id'], mod['id'])
-                        if mod['part_id'] not in deletes
-                        and mod['id'] not in fee_modifier_deletes else None)
-            for mod in fee_modifiers.values()}
-
-        # Check for duplicate fields in the same part.
+        # Check if each linked field and fee modifier name is unique.
+        used_fields: Set[int] = set()
+        used_names: Set[str] = set()
         field_msg = n_("Must not have multiple fee modifiers linked to the same"
                        " field in one event part.")
-        name_msg = n_("Must not have multiple fee modifiers witht he same name "
+        name_msg = n_("Must not have multiple fee modifiers with the same name "
                       "in one event part.")
-        used_fields = {}
-        used_names = {}
-        if len(ret_fee_modifiers) == 1:
-            f = unwrap(ret_fee_modifiers)
-            if f:
-                used_fields[f['part_id']] = {f['field_id']}
-                used_names[f['part_id']] = {f['modifier_name']}
-        for e1, e2 in itertools.combinations(
-                filter(None, ret_fee_modifiers.values()), 2):
-            used_fields.setdefault(e1['part_id'], set()).add(e1['field_id'])
-            used_fields.setdefault(e2['part_id'], set()).add(e2['field_id'])
-            used_names.setdefault(e1['part_id'], set()).add(e1['modifier_name'])
-            used_names.setdefault(e2['part_id'], set()).add(e2['modifier_name'])
-            if e1['part_id'] == e2['part_id']:
-                if e1['field_id'] == e2['field_id']:
-                    base_key = "fee_modifier_field_id_{}_{}"
-                    key1 = base_key.format(e1['part_id'], e1['id'])
-                    rs.add_validation_error((key1, ValueError(field_msg)))
-                    key2 = base_key.format(e2['part_id'], e2['id'])
-                    rs.add_validation_error((key2, ValueError(field_msg)))
-                if e1['modifier_name'] == e2['modifier_name']:
-                    base_key = "fee_modifier_modifier_name_{}_{}"
-                    key1 = base_key.format(e1['part_id'], e1['id'])
-                    rs.add_validation_error((key1, ValueError(name_msg)))
-                    key2 = base_key.format(e2['part_id'], e2['id'])
-                    rs.add_validation_error((key2, ValueError(name_msg)))
+        for anid, modifier in fee_modifier_data.items():
+            if modifier is None:
+                continue
+            if modifier['field_id'] in used_fields:
+                rs.append_validation_error(
+                    (f"{fee_modifier_prefix}field_id_{anid}", ValueError(field_msg))
+                )
+            if modifier['modifier_name'] in used_names:
+                rs.append_validation_error(
+                    (f"{fee_modifier_prefix}modifier_name_{anid}", ValueError(name_msg))
+                )
+            used_fields.add(modifier['field_id'])
+            used_names.add(modifier['modifier_name'])
 
-        for part_id in parts:
-            marker = 1
-            while marker < 2 ** 5:
-                will_create = unwrap(request_extractor(
-                    rs, {f"fee_modifier_create_{part_id}_-{marker}": bool}))
-                if will_create:
-                    if has_registrations:
-                        raise ValueError(n_(
-                            "Registrations exist, no creation."))
-                    params = fee_modifier_params(part_id, -marker)
-                    constraints = constraint_maker(part_id, -marker)
-                    new_fee_modifier = fee_modifier_excavator(
-                        request_extractor(rs, params, constraints),
-                        part_id, -marker)
-                    ret_fee_modifiers[-marker] = new_fee_modifier
-                    if new_fee_modifier['field_id'] in used_fields.get(
-                            part_id, set()):
-                        rs.add_validation_error(
-                            ("fee_modifier_field_id_{}_{}".format(
-                                part_id, -marker),
-                             ValueError(field_msg)))
-                    if new_fee_modifier['modifier_name'] in used_names.get(
-                            part_id, set()):
-                        rs.add_validation_error(
-                            ("fee_modifier_modifier_name_{}_{}".format(
-                                part_id, -marker),
-                             ValueError(name_msg)))
-                    used_fields.setdefault(part_id, set()).add(
-                        new_fee_modifier['field_id'])
-                else:
-                    break
-                marker += 1
-            rs.values['fee_modifier_create_last_index'][part_id] = marker - 1
-
-        # Don't allow fee modifiers for newly created parts.
-
-        # Handle deleted parts
-        for mod_id, mod in ret_fee_modifiers.items():
-            if mod:
-                ret[mod['part_id']].setdefault('fee_modifiers', {})[mod_id] = mod
-                del mod['part_id']
-                if 'id' in mod:
-                    del mod['id']
-        ret_parts = cast(Dict[int, Optional[CdEDBObject]], ret)
-        for part_id in deletes:
-            ret_parts[part_id] = None
-        if not any(ret_parts.values()):
-            rs.append_validation_error(
-                ("", ValueError(n_("At least one event part required."))))
-            rs.notify("error", n_("At least one event part required."))
-        return ret_parts
-
-    @access("event", modi={"POST"})
-    @event_guard(check_offline=True)
-    def part_summary(self, rs: RequestState, event_id: int) -> Response:
-        """Manipulate the parts of an event."""
-        has_registrations = self.eventproxy.has_registrations(rs, event_id)
-        parts = self.process_part_input(rs, has_registrations)
         if rs.has_validation_errors():
-            return self.part_summary_form(rs, event_id)
-        for part_id, part in rs.ambience['event']['parts'].items():
-            if parts.get(part_id) == part:
-                # remove unchanged
-                del parts[part_id]
+            return self.change_part_form(rs, event_id, part_id)
+
+        #
+        # put it all together
+        #
+        data['tracks'] = track_data
+        data['fee_modifiers'] = fee_modifier_data
         event = {
             'id': event_id,
-            'parts': parts,
+            'parts': {part_id: data},
         }
         code = self.eventproxy.set_event(rs, event)
         self.notify_return_code(rs, code)
-        return self.redirect(rs, "event/part_summary_form")
+
+        return self.redirect(rs, "event/part_summary")
 
     @access("event")
     @event_guard()
@@ -1021,52 +899,59 @@ class EventFrontend(AbstractUserFrontend):
                    if delete_flags['delete_{}'.format(field_id)]}
         ret: CdEDBOptionalMap = {}
 
-        def params_a(anid: int) -> TypeMapping:
+        def params_change(anid: int) -> TypeMapping:
+            """Return specification of parameters for changing an existing field."""
             return {
                 f"kind_{anid}": const.FieldDatatypes,
                 f"association_{anid}": const.FieldAssociations,
                 f"entries_{anid}": Optional[str],  # type: ignore
+                f"checkin_{anid}": bool,
             }
+        tmp: Optional[CdEDBObject]
         for field_id in fields:
             if field_id not in deletes:
-                tmp: Optional[CdEDBObject] = request_extractor(rs, params_a(field_id))
+                tmp = request_extractor(rs, params_change(field_id))
                 if rs.has_validation_errors():
                     break
-                tmp = check(rs, vtypes.EventField, tmp,
-                            extra_suffix="_{}".format(field_id))
+                tmp = check(rs, vtypes.EventField, tmp, extra_suffix=f"_{field_id}")
                 if tmp:
                     temp = {
-                        'kind': tmp["kind_{}".format(field_id)],
-                        'association': tmp["association_{}".format(field_id)],
-                        'entries': tmp["entries_{}".format(field_id)]}
+                        'kind': tmp[f"kind_{field_id}"],
+                        'association': tmp[f"association_{field_id}"],
+                        'entries': tmp[f"entries_{field_id}"],
+                        'checkin': tmp[f"checkin_{field_id}"],
+                    }
                     ret[field_id] = temp
         for field_id in deletes:
             ret[field_id] = None
         marker = 1
 
-        def params_b(anid: int) -> TypeMapping:
+        def params_creation(anid: int) -> TypeMapping:
+            """Return specification of parameters for creating a new field."""
             return {
                 f"field_name_-{anid}": str,
                 f"kind_-{anid}": const.FieldDatatypes,
                 f"association_-{anid}": const.FieldAssociations,
                 f"entries_-{anid}": Optional[str],  # type: ignore
+                f"checkin_-{anid}": bool,
             }
         while marker < 2 ** 10:
-            will_create = unwrap(request_extractor(
-                rs, {f"create_-{marker}": bool}))
+            will_create = unwrap(request_extractor(rs, {f"create_-{marker}": bool}))
             if will_create:
-                tmp = request_extractor(rs, params_b(marker))
+                tmp = request_extractor(rs, params_creation(marker))
                 if rs.has_validation_errors():
                     marker += 1
                     break
                 tmp = check(rs, vtypes.EventField, tmp, creation=True,
-                            extra_suffix="_-{}".format(marker))
+                            extra_suffix=f"_-{marker}")
                 if tmp:
                     temp = {
-                        'field_name': tmp["field_name_-{}".format(marker)],
-                        'kind': tmp["kind_-{}".format(marker)],
-                        'association': tmp["association_-{}".format(marker)],
-                        'entries': tmp["entries_-{}".format(marker)]}
+                        'field_name': tmp[f"field_name_-{marker}"],
+                        'kind': tmp[f"kind_-{marker}"],
+                        'association': tmp[f"association_-{marker}"],
+                        'entries': tmp[f"entries_-{marker}"],
+                        'checkin': tmp[f"checkin_-{marker}"],
+                    }
                     ret[-marker] = temp
             else:
                 break
@@ -1327,7 +1212,7 @@ class EventFrontend(AbstractUserFrontend):
         data['active_segments'] = active_segments
         field_params: TypeMapping = {
             f"fields.{field['field_name']}": Optional[  # type: ignore
-                VALIDATOR_LOOKUP[const.FieldDatatypes(field['kind']).name]]
+                VALIDATOR_LOOKUP[const.FieldDatatypes(field['kind']).name]]  # noqa: F821
             for field in rs.ambience['event']['fields'].values()
             if field['association'] == const.FieldAssociations.course
         }
@@ -1366,7 +1251,7 @@ class EventFrontend(AbstractUserFrontend):
         data['segments'] = segments
         field_params: TypeMapping = {
             f"fields.{field['field_name']}": Optional[  # type: ignore
-                VALIDATOR_LOOKUP[const.FieldDatatypes(field['kind']).name]]
+                VALIDATOR_LOOKUP[const.FieldDatatypes(field['kind']).name]]  # noqa: F821
             for field in rs.ambience['event']['fields'].values()
             if field['association'] == const.FieldAssociations.course
         }
@@ -1588,7 +1473,7 @@ class EventFrontend(AbstractUserFrontend):
         )
         QueryFilterGetter = Callable[
             [CdEDBObject, CdEDBObject, CdEDBObject], Collection[QueryConstraint]]
-        # Query filters for all the registration statistics defined and calculated above.
+        # Query filters for all the registration statistics defined and calculated above
         # They are customized and inserted into the query on the fly by get_query().
         # `e` is the event, `p` is the event_part, `t` is the track.
         registration_query_filters: Dict[str, QueryFilterGetter] = {
@@ -1630,6 +1515,7 @@ class EventFrontend(AbstractUserFrontend):
                 ('persona.id', QueryOperators.oneof,
                  rs.ambience['event']['orgas']),),
             'waitlist': lambda e, p, t: (
+                involved_filter(p),
                 ('part{}.status'.format(p['id']), QueryOperators.equal,
                  stati.waitlist.value),),
             'guest': lambda e, p, t: (
@@ -2046,13 +1932,11 @@ class EventFrontend(AbstractUserFrontend):
                 if assign_action.enum == CourseChoiceToolActions.specific_rank:
                     if assign_action.int >= len(reg_track['choices']):
                         rs.notify("warning",
-                                  (n_("%(given_names)s %(family_name)s has no "
+                                  (n_("%(name)s has no "
                                       "%(rank)i. choice in %(track_name)s.")
                                    if len(tracks) > 1
-                                   else n_("%(given_names)s %(family_name)s "
-                                           "has no %(rank)i. choice.")),
-                                  {'given_names': persona['given_names'],
-                                   'family_name': persona['family_name'],
+                                   else n_("%(name)s has no %(rank)i. choice.")),
+                                  {'name': make_persona_name(persona),
                                    'rank': assign_action.int + 1,
                                    'track_name': tracks[atrack_id]['title']})
                         continue
@@ -2082,13 +1966,12 @@ class EventFrontend(AbstractUserFrontend):
                             break
                     else:
                         rs.notify("warning",
-                                  (n_("No choice available for %(given_names)s "
-                                      "%(family_name)s in %(track_name)s.")
+                                  (n_("No choice available for %(name)s in "
+                                      "%(track_name)s.")
                                    if len(tracks) > 1
                                    else n_("No choice available for "
-                                           "%(given_names)s %(family_name)s.")),
-                                  {'given_names': persona['given_names'],
-                                   'family_name': persona['family_name'],
+                                           "%(name)s.")),
+                                  {'name': make_persona_name(persona),
                                    'track_name': tracks[atrack_id]['title']})
             if tmp['tracks']:
                 res = self.eventproxy.set_registration(rs, tmp, change_note)
@@ -2096,10 +1979,8 @@ class EventFrontend(AbstractUserFrontend):
                     num_committed += 1
                 else:
                     rs.notify("warning",
-                              n_("Error committing changes for %(given_names)s "
-                                 "%(family_name)s."),
-                              {'given_names': persona['given_names'],
-                               'family_name': persona['family_name']})
+                              n_("Error committing changes for %(name)s."),
+                              {'name': make_persona_name(persona)})
         rs.notify("success" if num_committed > 0 else "warning",
                   n_("Course assignment for %(num_committed)s of %(num_total)s "
                      "registrations committed."),
@@ -2294,8 +2175,8 @@ class EventFrontend(AbstractUserFrontend):
         })
         return datum
 
-    def book_fees(self, rs: RequestState, data: Collection[CdEDBObject],
-                  send_notifications: bool = False
+    def book_fees(self, rs: RequestState,
+                  data: Collection[CdEDBObject], send_notifications: bool = False
                   ) -> Tuple[bool, Optional[int]]:
         """Book all paid fees.
 
@@ -2305,56 +2186,27 @@ class EventFrontend(AbstractUserFrontend):
           * for negative outcome the line where an exception was triggered
             or None if it was a DB serialization error
         """
-        index = 0
-        try:
-            with Atomizer(rs):
-                count = 0
-                all_reg_ids = {datum['registration_id'] for datum in data}
-                all_regs = self.eventproxy.get_registrations(rs, all_reg_ids)
-                for index, datum in enumerate(data):
-                    reg_id = datum['registration_id']
-                    update = {
-                        'id': reg_id,
-                        'payment': datum['date'],
-                        'amount_paid': all_regs[reg_id]['amount_paid']
-                                       + datum['amount'],
+        relevant_keys = {'registration_id', 'date', 'original_date', 'amount'}
+        relevant_data = [{k: v for k, v in item.items() if k in relevant_keys}
+                         for item in data]
+        with TransactionObserver(rs, self, "book_fees"):
+            success, number = self.eventproxy.book_fees(
+                rs, rs.ambience['event']['id'], relevant_data)
+            if success and send_notifications:
+                persona_ids = tuple(e['persona_id'] for e in data)
+                personas = self.coreproxy.get_personas(rs, persona_ids)
+                subject = "Überweisung für {} eingetroffen".format(
+                    rs.ambience['event']['title'])
+                for persona in personas.values():
+                    headers: Dict[str, Union[str, Collection[str]]] = {
+                        'To': (persona['username'],),
+                        'Subject': subject,
                     }
-                    change_note = "{} am {} gezahlt.".format(
-                        money_filter(datum['amount']),
-                        date_filter(datum['original_date'], lang="de"))
-                    count += self.eventproxy.set_registration(rs, update, change_note)
-        except psycopg2.extensions.TransactionRollbackError:
-            # We perform a rather big transaction, so serialization errors
-            # could happen.
-            return False, None
-        except Exception:
-            # This blanket catching of all exceptions is a last resort. We try
-            # to do enough validation, so that this should never happen, but
-            # an opaque error (as would happen without this) would be rather
-            # frustrating for the users -- hence some extra error handling
-            # here.
-            self.logger.error(glue(
-                ">>>\n>>>\n>>>\n>>> Exception during fee transfer processing",
-                "<<<\n<<<\n<<<\n<<<"))
-            self.logger.exception("FIRST AS SIMPLE TRACEBACK")
-            self.logger.error("SECOND TRY CGITB")
-            self.cgitb_log()
-            return False, index
-        if send_notifications:
-            persona_ids = tuple(e['persona_id'] for e in data)
-            personas = self.coreproxy.get_personas(rs, persona_ids)
-            subject = "Überweisung für {} eingetroffen".format(
-                rs.ambience['event']['title'])
-            for persona in personas.values():
-                headers: Dict[str, Union[str, Collection[str]]] = {
-                    'To': (persona['username'],),
-                    'Subject': subject,
-                }
-                if rs.ambience['event']['orga_address']:
-                    headers['Reply-To'] = rs.ambience['event']['orga_address']
-                self.do_mail(rs, "transfer_received", headers,
-                             {'persona': persona})
-        return True, count
+                    if rs.ambience['event']['orga_address']:
+                        headers['Reply-To'] = rs.ambience['event']['orga_address']
+                    self.do_mail(rs, "transfer_received", headers,
+                                 {'persona': persona})
+            return success, number
 
     @access("event", modi={"POST"})
     @event_guard(check_offline=True)
@@ -2473,7 +2325,7 @@ class EventFrontend(AbstractUserFrontend):
             work_dir.mkdir()
             filename = "{}_nametags.tex".format(
                 rs.ambience['event']['shortname'])
-            with open(work_dir / filename, 'w') as f:
+            with open(work_dir / filename, 'w', encoding='utf-8') as f:
                 f.write(tex)
             src = self.conf["REPOSITORY_PATH"] / "misc/blank.png"
             shutil.copy(src, work_dir / "aka-logo.png")
@@ -2574,34 +2426,30 @@ class EventFrontend(AbstractUserFrontend):
         lodgement_ids = self.eventproxy.list_lodgements(rs, event_id)
         lodgements = self.eventproxy.get_lodgements(rs, lodgement_ids)
 
-        reverse_wish = {}
+        rwish = collections.defaultdict(list)
         if event['lodge_field']:
-            for reg_id, reg in registrations.items():
-                rwish = set()
-                persona = personas[reg['persona_id']]
-                checks = {
-                    diacritic_patterns("{} {}".format(
-                        given_name, persona['family_name']))
-                    for given_name in persona['given_names'].split()}
-                checks.add(diacritic_patterns("{} {}".format(
-                    persona['display_name'], persona['family_name'])))
-                for oid, other in registrations.items():
-                    owish = other['fields'].get(
-                        event['fields'][event['lodge_field']]['field_name'])
-                    if not owish:
-                        continue
-                    if any(re.search(acheck, owish, flags=re.IGNORECASE)
-                                     for acheck in checks):
-                        rwish.add(oid)
-                reverse_wish[reg_id] = ", ".join(
-                    "{} {}".format(
-                        personas[registrations[id]['persona_id']]['given_names'],
-                        personas[registrations[id]['persona_id']]['family_name'])
-                    for id in rwish)
+            wishes, problems = detect_lodgement_wishes(registrations, personas,
+                                                       event, None)
+            for wish in wishes:
+                if wish.negated:
+                    continue
+                rwish[wish.wished].append(wish.wishing)
+                if wish.bidirectional:
+                    rwish[wish.wishing].append(wish.wished)
+        else:
+            problems = []
+        reverse_wish = {
+            reg_id: ", ".join(
+                make_persona_name(
+                    personas[registrations[wishing_id]['persona_id']])
+                for wishing_id in rwish[reg_id])
+            for reg_id in registrations
+        }
 
         tex = self.fill_template(rs, "tex", "lodgement_puzzle", {
             'lodgements': lodgements, 'registrations': registrations,
-            'personas': personas, 'reverse_wish': reverse_wish})
+            'personas': personas, 'reverse_wish': reverse_wish,
+            'wish_problems': problems})
         file = self.serve_latex_document(rs, tex, "{}_lodgement_puzzle".format(
             rs.ambience['event']['shortname']), runs)
         if file:
@@ -2662,7 +2510,7 @@ class EventFrontend(AbstractUserFrontend):
             work_dir.mkdir()
             filename = "{}_course_lists.tex".format(
                 rs.ambience['event']['shortname'])
-            with open(work_dir / filename, 'w') as f:
+            with open(work_dir / filename, 'w', encoding='utf-8') as f:
                 f.write(tex)
             src = self.conf["REPOSITORY_PATH"] / "misc/blank.png"
             shutil.copy(src, work_dir / "event-logo.png")
@@ -2708,7 +2556,7 @@ class EventFrontend(AbstractUserFrontend):
             work_dir.mkdir()
             filename = "{}_lodgement_lists.tex".format(
                 rs.ambience['event']['shortname'])
-            with open(work_dir / filename, 'w') as f:
+            with open(work_dir / filename, 'w', encoding='utf-8') as f:
                 f.write(tex)
             src = self.conf["REPOSITORY_PATH"] / "misc/blank.png"
             shutil.copy(src, work_dir / "aka-logo.png")
@@ -2759,8 +2607,11 @@ class EventFrontend(AbstractUserFrontend):
             rs.notify("info", n_("Empty File."))
             return self.redirect(rs, "event/downloads")
         courses = self.eventproxy.get_courses(rs, course_ids)
-        data = self.fill_template(
-            rs, "other", "dokuteam_courselist", {'courses': courses})
+        active_courses = filter(lambda c: c["active_segments"], courses.values())
+        sorted_courses = xsorted(active_courses, key=EntitySorter.course)
+        data = self.fill_template(rs, "other", "dokuteam_courselist", {
+            "sorted_courses": sorted_courses
+        })
         return self.send_file(
             rs, data=data, inline=False,
             filename=f"{rs.ambience['event']['shortname']}_dokuteam_courselist.txt")
@@ -2789,7 +2640,8 @@ class EventFrontend(AbstractUserFrontend):
                     order = [("persona.given_names", True)]
                     query = Query(QueryScope.registration, spec, fields_of_interest,
                                   constrains, order)
-                    query_res = self.eventproxy.submit_general_query(rs, query, event_id)
+                    query_res = self.eventproxy.submit_general_query(rs, query,
+                                                                     event_id)
                     course_key = f"track{track_id}.course_id"
                     # we have to replace the course id with the course number
                     result = tuple(
@@ -2806,7 +2658,7 @@ class EventFrontend(AbstractUserFrontend):
                     # save the result in one file per track
                     filename = f"{asciificator(track['shortname'])}.csv"
                     file = pathlib.Path(work_dir, filename)
-                    file.write_text(data)
+                    file.write_text(data, encoding='utf-8')
 
             # create a zip archive of all lists
             zipname = f"{rs.ambience['event']['shortname']}_dokuteam_participant_list"
@@ -2947,6 +2799,48 @@ class EventFrontend(AbstractUserFrontend):
         ret['export'] = self.eventproxy.partial_export_event(rs, event_id)
         ret['message'] = "success"
         return self.send_json(rs, ret)
+
+    @access("event")
+    @event_guard()
+    def questionnaire_import_form(self, rs: RequestState, event_id: int) -> Response:
+        """Render form for uploading questionnaire data."""
+        return self.render(rs, "questionnaire_import")
+
+    @access("event", modi={"POST"})
+    @event_guard(check_offline=True)
+    @REQUESTfile("json_file")
+    @REQUESTdata("extend_questionnaire", "skip_existing_fields", "token")
+    def questionnaire_import(
+        self, rs: RequestState, event_id: int,
+        json_file: Optional[werkzeug.datastructures.FileStorage],
+        extend_questionnaire: bool, skip_existing_fields: bool, token: Optional[str],
+    ) -> Response:
+        """Import questionnaire rows and custom datafields.
+
+        :param extend_questionnaire: If True, append the imported questionnaire rows to
+            any existing ones. Otherwise replace the existing ones.
+        :param skip_existing_fields: If True, the import of fields that already exist
+            is skipped, even if their definition is different from the existing one.
+            Otherwise, duplicate field names will cause an error and prevent the import.
+        """
+        kwargs = {
+            'field_definitions': rs.ambience['event']['fields'],
+            'fee_modifiers': rs.ambience['event']['fee_modifiers'],
+            'questionnaire': self.eventproxy.get_questionnaire(rs, event_id),
+            'extend_questionnaire': extend_questionnaire,
+            'skip_existing_fields': skip_existing_fields,
+        }
+        data = check(rs, vtypes.SerializedEventQuestionnaireUpload, json_file,
+                     **kwargs)
+        if rs.has_validation_errors():
+            return self.questionnaire_import_form(rs, event_id)
+        assert data is not None
+
+        code = self.eventproxy.questionnaire_import(
+            rs, event_id, fields=data['fields'], questionnaire=data['questionnaire'])
+
+        self.notify_return_code(rs, code)
+        return self.redirect(rs, "event/configure_additional_questionnaire")
 
     @access("event")
     @event_guard()
@@ -3187,8 +3081,7 @@ class EventFrontend(AbstractUserFrontend):
         course_entries = {
             c["id"]: "{}. {}".format(c["nr"], c["shortname"])
             for c in courses.values()}
-        lodgement_entries = {l["id"]: l["title"]
-                             for l in lodgements.values()}
+        lodgement_entries = {lgd["id"]: lgd["title"] for lgd in lodgements.values()}
         reg_part_stati_entries =\
             dict(enum_entries_filter(const.RegistrationPartStati, rs.gettext))
         segment_stati_entries = {
@@ -3773,7 +3666,7 @@ class EventFrontend(AbstractUserFrontend):
         f = lambda entry: rs.ambience['event']['fields'][entry['field_id']]
         params: TypeMapping = {
             f(entry)['field_name']: Optional[  # type: ignore
-                VALIDATOR_LOOKUP[const.FieldDatatypes(f(entry)['kind']).name]]
+                VALIDATOR_LOOKUP[const.FieldDatatypes(f(entry)['kind']).name]]  # noqa: F821
             for entry in add_questionnaire
             if entry['field_id'] and not entry['readonly']
         }
@@ -3959,7 +3852,8 @@ class EventFrontend(AbstractUserFrontend):
 
         new_questionnaire = [self._sanitize_questionnaire_row(questionnaire[i])
                              for i in order]
-        code = self.eventproxy.set_questionnaire(rs, event_id, {kind: new_questionnaire})
+        code = self.eventproxy.set_questionnaire(rs, event_id,
+                                                 {kind: new_questionnaire})
         self.notify_return_code(rs, code)
         return self.redirect(rs, "event/reorder_questionnaire_form", {'kind': kind})
 
@@ -4115,7 +4009,7 @@ class EventFrontend(AbstractUserFrontend):
             })
         field_params: TypeMapping = {
             f"fields.{field['field_name']}": Optional[  # type: ignore
-                VALIDATOR_LOOKUP[const.FieldDatatypes(field['kind']).name]]
+                VALIDATOR_LOOKUP[const.FieldDatatypes(field['kind']).name]]  # noqa: F821
             for field in event['fields'].values()
             if field['association'] == const.FieldAssociations.registration
         }
@@ -4766,8 +4660,9 @@ class EventFrontend(AbstractUserFrontend):
                                 ) -> Response:
         """Manipulate groups of lodgements."""
         group_ids = self.eventproxy.list_lodgement_groups(rs, event_id)
-        groups = process_dynamic_input(rs, group_ids.keys(), {'title': str},
-                                       {'event_id': event_id})
+        spec = {'title': str}
+        groups = process_dynamic_input(
+            rs, group_ids.keys(), spec, additional={'event_id': event_id})
         if rs.has_validation_errors():
             return self.lodgement_group_summary_form(rs, event_id)
         code = 1
@@ -4778,11 +4673,8 @@ class EventFrontend(AbstractUserFrontend):
             elif group_id < 0:
                 code *= self.eventproxy.create_lodgement_group(rs, group)
             else:
-                with Atomizer(rs):
-                    current = self.eventproxy.get_lodgement_group(rs, group_id)
-                    # Do not update unchanged
-                    if current != group:
-                        code *= self.eventproxy.set_lodgement_group(rs, group)
+                del group['event_id']
+                code *= self.eventproxy.rcw_lodgement_group(rs, group)
         self.notify_return_code(rs, code)
         return self.redirect(rs, "event/lodgement_group_summary")
 
@@ -4821,6 +4713,60 @@ class EventFrontend(AbstractUserFrontend):
         })
 
     @access("event")
+    @event_guard()
+    def lodgement_wishes_graph_form(self, rs: RequestState, event_id: int
+                                    ) -> Response:
+        event = rs.ambience['event']
+        if event['lodge_field']:
+            registration_ids = self.eventproxy.list_registrations(rs, event_id)
+            registrations = self.eventproxy.get_registrations(rs, registration_ids)
+            personas = self.coreproxy.get_event_users(rs, tuple(
+                reg['persona_id'] for reg in registrations.values()), event_id)
+
+            _wishes, problems = detect_lodgement_wishes(
+                registrations, personas, event, restrict_part_id=None)
+        else:
+            problems = []
+        return self.render(rs, "lodgement_wishes_graph_form",
+                           {'problems': problems})
+
+    @access("event")
+    @event_guard()
+    @REQUESTdata('all_participants', 'part_id', 'show_lodgements')
+    def lodgement_wishes_graph(self, rs: RequestState, event_id: int,
+                               all_participants: bool, part_id: Optional[int],
+                               show_lodgements: bool) -> Response:
+        if rs.has_validation_errors():
+            return self.redirect(rs, 'event/lodgement_wishes_graph_form')
+        event = rs.ambience['event']
+
+        if not event['lodge_field']:
+            rs.notify('error', n_("Lodgement wishes graph is only available if "
+                                  "the Field for Rooming Preferences is set in "
+                                  "event configuration."))
+            return self.redirect(rs, 'event/lodgement_wishes_graph_form')
+        if show_lodgements and not part_id:
+            rs.notify('error', n_("Lodgement clusters can only be displayed if "
+                                  "the graph is restricted to a specific "
+                                  "part."))
+            return self.redirect(rs, 'event/lodgement_wishes_graph_form')
+
+        registration_ids = self.eventproxy.list_registrations(rs, event_id)
+        registrations = self.eventproxy.get_registrations(rs, registration_ids)
+        lodgement_ids = self.eventproxy.list_lodgements(rs, event_id)
+        lodgements = self.eventproxy.get_lodgements(rs, lodgement_ids)
+        personas = self.coreproxy.get_event_users(rs, tuple(
+            reg['persona_id'] for reg in registrations.values()), event_id)
+
+        wishes, _problems = detect_lodgement_wishes(
+            registrations, personas, event, part_id)
+        graph = create_lodgement_wishes_graph(
+            rs, registrations, wishes, lodgements, event, personas, part_id,
+            all_participants, part_id if show_lodgements else None)
+        data: bytes = graph.pipe('svg')
+        return self.send_file(rs, "image/svg+xml", data=data)
+
+    @access("event")
     @event_guard(check_offline=True)
     def create_lodgement_form(self, rs: RequestState, event_id: int
                               ) -> Response:
@@ -4837,7 +4783,7 @@ class EventFrontend(AbstractUserFrontend):
         data['event_id'] = event_id
         field_params: TypeMapping = {
             f"fields.{field['field_name']}": Optional[  # type: ignore
-                VALIDATOR_LOOKUP[const.FieldDatatypes(field['kind']).name]]
+                VALIDATOR_LOOKUP[const.FieldDatatypes(field['kind']).name]]  # noqa: F821
             for field in rs.ambience['event']['fields'].values()
             if field['association'] == const.FieldAssociations.lodgement
         }
@@ -4879,7 +4825,7 @@ class EventFrontend(AbstractUserFrontend):
         data['id'] = lodgement_id
         field_params: TypeMapping = {
             f"fields.{field['field_name']}": Optional[  # type: ignore
-                VALIDATOR_LOOKUP[const.FieldDatatypes(field['kind']).name]]
+                VALIDATOR_LOOKUP[const.FieldDatatypes(field['kind']).name]]  # noqa: F821
             for field in rs.ambience['event']['fields'].values()
             if field['association'] == const.FieldAssociations.lodgement
         }
@@ -4967,7 +4913,8 @@ class EventFrontend(AbstractUserFrontend):
                  if _check_not_this_lodgement(registration_id, part_id)],
                 key=lambda x: (
                     x['current'] is not None,
-                    EntitySorter.persona(personas[registrations[x['id']]['persona_id']]))
+                    EntitySorter.persona(
+                        personas[registrations[x['id']]['persona_id']]))
             )
             for part_id in rs.ambience['event']['parts']
         }
@@ -5073,6 +5020,7 @@ class EventFrontend(AbstractUserFrontend):
             lodgements.keys(), rs.ambience['event'], registrations, key="lodgement_id")
 
         new_regs: CdEDBObjectMap = {}
+        change_notes = []
         for part_id in rs.ambience['event']['parts']:
             if data[f"swap_with_{part_id}"]:
                 swap_lodgement_id: int = data[f"swap_with_{part_id}"]
@@ -5087,10 +5035,13 @@ class EventFrontend(AbstractUserFrontend):
                     new_reg = new_regs.get(reg_id, {'id': reg_id, 'parts': dict()})
                     new_reg['parts'][part_id] = {'lodgement_id': lodgement_id}
                     new_regs[reg_id] = new_reg
+                change_notes.append(
+                    f"Bewohner von {lodgements[lodgement_id]} und"
+                    f" {lodgements[swap_lodgement_id]} für"
+                    f" {rs.ambience['event']['parts'][part_id]['title']} getauscht")
 
         code = 1
-        change_note = (f"Bewohner von {lodgements[lodgement_id]} und"
-                       f" {lodgements[swap_lodgement_id]} getauscht.")
+        change_note = ", ".join(change_notes) + "."
         for new_reg in new_regs.values():
             code *= self.eventproxy.set_registration(rs, new_reg, change_note)
         self.notify_return_code(rs, code)
@@ -5142,13 +5093,11 @@ class EventFrontend(AbstractUserFrontend):
 
         selectize_data = {
             track_id: xsorted(
-                [{'name': (personas[registration['persona_id']]['given_names']
-                           + " " + personas[registration['persona_id']]
-                           ['family_name']),
+                ({'name': make_persona_name(personas[registration['persona_id']]),
                   'current': registration['tracks'][track_id]['course_id'],
                   'id': registration_id}
                  for registration_id, registration in registrations.items()
-                 if _check_not_this_course(registration_id, track_id)],
+                 if _check_not_this_course(registration_id, track_id)),
                 key=lambda x: (
                     x['current'] is not None,
                     EntitySorter.persona(
@@ -5465,17 +5414,21 @@ class EventFrontend(AbstractUserFrontend):
 
     @access("event")
     @event_guard(check_offline=True)
-    def checkin_form(self, rs: RequestState, event_id: int) -> Response:
+    @REQUESTdata("part_ids")
+    def checkin_form(self, rs: RequestState, event_id: int,
+                     part_ids: Optional[vtypes.IntCSVList] = None) -> Response:
         """Render form."""
+        if rs.has_validation_errors() or not part_ids:
+            parts = rs.ambience['event']['parts']
+        else:
+            parts = {p_id: rs.ambience['event']['parts'][p_id] for p_id in part_ids}
         registration_ids = self.eventproxy.list_registrations(rs, event_id)
         registrations = self.eventproxy.get_registrations(rs, registration_ids)
         there = lambda registration, part_id: const.RegistrationPartStati(
             registration['parts'][part_id]['status']).is_present()
         registrations = {
-            k: v
-            for k, v in registrations.items()
-            if (not v['checkin']
-                and any(there(v, id) for id in rs.ambience['event']['parts']))}
+            k: v for k, v in registrations.items()
+            if (not v['checkin'] and any(there(v, id) for id in parts))}
         personas = self.coreproxy.get_event_users(rs, tuple(
             reg['persona_id'] for reg in registrations.values()), event_id)
         lodgement_ids = self.eventproxy.list_lodgements(rs, event_id)
@@ -5490,15 +5443,21 @@ class EventFrontend(AbstractUserFrontend):
                 personas[registrations[anid]['persona_id']]))
         registrations = OrderedDict(
             (reg_id, registrations[reg_id]) for reg_id in reg_order)
+        checkin_fields = {
+            field_id: f for field_id, f in rs.ambience['event']['fields'].items()
+            if f['checkin'] and f['association'] == const.FieldAssociations.registration
+        }
         return self.render(rs, "checkin", {
             'registrations': registrations, 'personas': personas,
-            'lodgements': lodgements})
+            'lodgements': lodgements, 'checkin_fields': checkin_fields,
+            'part_ids_param': ",".join(map(str, parts))
+        })
 
     @access("event", modi={"POST"})
     @event_guard(check_offline=True)
-    @REQUESTdata("registration_id")
-    def checkin(self, rs: RequestState, event_id: int, registration_id: vtypes.ID
-                ) -> Response:
+    @REQUESTdata("registration_id", "part_ids")
+    def checkin(self, rs: RequestState, event_id: int, registration_id: vtypes.ID,
+                part_ids: Optional[vtypes.IntCSVList] = None) -> Response:
         """Check a participant in."""
         if rs.has_validation_errors():
             return self.checkin_form(rs, event_id)
@@ -5515,7 +5474,8 @@ class EventFrontend(AbstractUserFrontend):
         }
         code = self.eventproxy.set_registration(rs, new_reg, "Eingecheckt.")
         self.notify_return_code(rs, code)
-        return self.redirect(rs, 'event/checkin')
+        params = {'part_ids': ",".join(map(str, part_ids))} if part_ids else None
+        return self.redirect(rs, 'event/checkin', params)
 
     FIELD_REDIRECT = {
         const.FieldAssociations.registration: "event/registration_query",
@@ -5535,10 +5495,12 @@ class EventFrontend(AbstractUserFrontend):
         :param kind: specifies the entity: registration, course or lodgement
 
         :returns: A tuple of values, containing
-            * entities: corresponding to the given ids (registrations, courses, lodgements)
+            * entities: corresponding to the given ids (registrations, courses,
+                lodgements)
             * ordered_ids: given ids, sorted by the corresponding EntitySorter
             * labels: name of the entities which will be displayed in the template
-            * field: the event field which will be changed, None if no field_id was given
+            * field: the event field which will be changed, None if no field_id was
+                given
         """
         if kind == const.FieldAssociations.registration:
             if not ids:
@@ -5547,8 +5509,7 @@ class EventFrontend(AbstractUserFrontend):
             personas = self.coreproxy.get_personas(
                 rs, tuple(e['persona_id'] for e in entities.values()))
             labels = {
-                reg_id: (f"{personas[entity['persona_id']]['given_names']}"
-                         f" {personas[entity['persona_id']]['family_name']}")
+                reg_id: make_persona_name(personas[entity['persona_id']])
                 for reg_id, entity in entities.items()}
             ordered_ids = xsorted(
                 entities.keys(), key=lambda anid: EntitySorter.persona(
@@ -5570,11 +5531,12 @@ class EventFrontend(AbstractUserFrontend):
             groups = self.eventproxy.get_lodgement_groups(rs, group_ids)
             labels = {
                 lodg_id: f"{lodg['title']}" if lodg['group_id'] is None
-                         else safe_filter(f"{lodg['title']}, "
-                                          f"<em>{groups[lodg['group_id']]['title']}</em>")
+                         else safe_filter(f"{lodg['title']}, <em>"
+                                          f"{groups[lodg['group_id']]['title']}</em>")
                 for lodg_id, lodg in entities.items()}
             ordered_ids = xsorted(
-                entities.keys(), key=lambda anid: EntitySorter.lodgement(entities[anid]))
+                entities.keys(),
+                key=lambda anid: EntitySorter.lodgement(entities[anid]))
         else:
             # this should not happen, since we check before for validation errors
             raise NotImplementedError(f"Unknown kind {kind}")
@@ -5608,7 +5570,8 @@ class EventFrontend(AbstractUserFrontend):
                 rs, "event/field_set_form", {
                     'ids': (','.join(str(i) for i in ids) if ids else None),
                     'field_id': field_id, 'kind': kind.value})
-        _, ordered_ids, labels, _ = self.field_set_aux(rs, event_id, field_id, ids, kind)
+        _, ordered_ids, labels, _ = self.field_set_aux(rs, event_id, field_id, ids,
+                                                       kind)
         fields = [(field['id'], field['field_name'])
                   for field in xsorted(rs.ambience['event']['fields'].values(),
                                        key=EntitySorter.event_field)
