@@ -385,7 +385,7 @@ CDEDB_PATHS = werkzeug.routing.Map((
                 rule("/ml/create", methods=_POST,
                      endpoint="create_event_mailinglist"),
                 rule("/part/summary", methods=_GET,
-                     endpoint="part_summary_form"),
+                     endpoint="part_summary"),
                 rule("/part/summary", methods=_POST,
                      endpoint="part_summary"),
                 rule("/course/list", methods=_GET,
@@ -450,6 +450,10 @@ CDEDB_PATHS = werkzeug.routing.Map((
                      endpoint="lodgement_group_summary"),
                 rule("/lodgement/query", methods=_GET,
                      endpoint="lodgement_query"),
+                rule("/lodgement/graph/show", methods=_GET,
+                     endpoint="lodgement_wishes_graph"),
+                rule("/lodgement/graph/form", methods=_GET,
+                     endpoint="lodgement_wishes_graph_form"),
                 rule("/registration/query", methods=_GET,
                      endpoint="registration_query"),
                 rule("/checkin", methods=_GET,
@@ -468,10 +472,15 @@ CDEDB_PATHS = werkzeug.routing.Map((
                      endpoint="delete_event_query"),
                 rule("/download", methods=_GET,
                      endpoint="downloads"),
-                rule("/import", methods=_GET,
-                     endpoint="partial_import_form"),
-                rule("/import", methods=_POST,
-                     endpoint="partial_import"),
+                sub("/import", (
+                    rule("/partial", methods=_GET,
+                         endpoint="partial_import_form"),
+                    rule("/partial", methods=_POST,
+                         endpoint="partial_import"),
+                    rule("/questionnaire", methods=_GET,
+                         endpoint="questionnaire_import_form"),
+                    rule("/questionnaire", methods=_POST,
+                         endpoint="questionnaire_import"),)),
                 rule("/archive", methods=_POST,
                      endpoint="archive_event"),
                 rule("/delete", methods=_POST,
@@ -501,6 +510,17 @@ CDEDB_PATHS = werkzeug.routing.Map((
                          endpoint="download_csv_lodgements"),
                     rule("/csv_registrations", methods=_GET,
                          endpoint="download_csv_registrations"),)),
+                rule("/part/add", methods=_GET,
+                     endpoint="add_part_form"),
+                rule("/part/add", methods=_POST,
+                     endpoint="add_part"),
+                sub('/part/<int:part_id>', (
+                    rule("/change", methods=_GET,
+                         endpoint="change_part_form"),
+                    rule("/change", methods=_POST,
+                         endpoint="change_part"),
+                    rule("/delete", methods=_POST,
+                         endpoint="delete_part"),)),
                 sub('/course/<int:course_id>', (
                     rule("/show", methods=_GET,
                          endpoint="show_course"),
@@ -617,29 +637,28 @@ CDEDB_PATHS = werkzeug.routing.Map((
                     rule("/add", methods=_POST,
                          endpoint="add_attachment"),
                     sub("/<int:attachment_id>", (
-                        rule("/get", methods=_GET,
+                        rule("/latest", methods=_GET,
                              endpoint="get_attachment"),
-                        rule("/show", methods=_GET,
-                             endpoint="show_attachment"),
-                        rule("/change", methods=_GET,
-                             endpoint="change_attachment_link_form"),
-                        rule("/change", methods=_POST,
-                             endpoint="change_attachment_link"),
+                        # this is kept as legacy support
+                        rule("/get", endpoint="",
+                             redirect_to='assembly/assembly/<assembly_id>/attachment'
+                                         '/<attachment_id>/latest'),
                         rule("/delete", methods=_POST,
                              endpoint="delete_attachment"),
                         rule("/add", methods=_GET,
-                             endpoint="add_attachment_form"),
+                             endpoint="add_attachment_version_form"),
                         rule("/add", methods=_POST,
-                             endpoint="add_attachment"),
-                        sub("/version/<int:version>", (
-                            rule("/get", methods=_GET,
-                                 endpoint="get_attachment"),
+                             endpoint="add_attachment_version"),
+                        sub("/version/<int:version_nr>", (
+                            rule("/", methods=_GET,
+                                 endpoint="get_attachment_version"),
+                            # this is kept as legacy support
+                            rule("/get", endpoint="",
+                                 redirect_to="assembly/assembly/<assembly_id>"
+                                             "/attachment/<attachment_id>/version"
+                                             "/<version_nr>"),
                             rule("/delete", methods=_POST,
-                                 endpoint="delete_attachment"),
-                            rule("/edit", methods=_GET,
-                                 endpoint="edit_attachment_version_form"),
-                            rule("/edit", methods=_POST,
-                                 endpoint="edit_attachment_version"),
+                                 endpoint="delete_attachment_version"),
                         )),
                     )),
                 )),
@@ -671,34 +690,17 @@ CDEDB_PATHS = werkzeug.routing.Map((
                     rule("/result/download", methods=_GET,
                          endpoint="get_result"),
                     sub("/attachment", (
-                        rule("/add", methods=_GET,
-                             endpoint="add_attachment_form"),
-                        rule("/add", methods=_POST,
-                             endpoint="add_attachment"),
                         sub("/<int:attachment_id>", (
-                            rule("/get", methods=_GET,
-                                 endpoint="get_attachment"),
-                            rule("/show", methods=_GET,
-                                 endpoint="show_attachment"),
-                            rule("/change", methods=_GET,
-                                 endpoint="change_attachment_link_form"),
-                            rule("/change", methods=_POST,
-                                 endpoint="change_attachment_link"),
-                            rule("/delete", methods=_POST,
-                                 endpoint="delete_attachment"),
-                            rule("/add", methods=_GET,
-                                 endpoint="add_attachment_form"),
-                            rule("/add", methods=_POST,
-                                 endpoint="add_attachment"),
+                            # this is kept as legacy support
+                            rule("/get", endpoint="",
+                                 redirect_to='assembly/assembly/<assembly_id>'
+                                             '/attachment/<attachment_id>/latest'),
                             sub("/version/<int:version>", (
-                                rule("/get", methods=_GET,
-                                     endpoint="get_attachment"),
-                                rule("/delete", methods=_POST,
-                                     endpoint="delete_attachment"),
-                                rule("/edit", methods=_GET,
-                                     endpoint="edit_attachment_version_form"),
-                                rule("/edit", methods=_POST,
-                                     endpoint="edit_attachment_version"),
+                                # this is kept as legacy support
+                                rule("/get", endpoint="",
+                                     redirect_to="assembly/assembly/<assembly_id>"
+                                                 "/attachment/<attachment_id>/version"
+                                                 "/<version>"),
                             )),
                         )),
                     )),
@@ -759,6 +761,8 @@ CDEDB_PATHS = werkzeug.routing.Map((
                      endpoint="message_moderation_form"),
                 rule("/moderate", methods=_POST,
                      endpoint="message_moderation"),
+                rule("/moderate/multi", methods=_POST,
+                     endpoint="message_moderation_multi"),
                 rule("/management/advanced", methods=_GET,
                      endpoint="advanced_management"),
                 rule("/download/subscriptions", methods=_GET,
