@@ -1,11 +1,13 @@
+"""Classes for different types of mailinglists"""
+
 import enum
 import itertools
+import typing
 from collections import OrderedDict
 from typing import (
-    TYPE_CHECKING, Any, Collection, Dict, List, Mapping, Optional, Set, Type, Union,
+    TYPE_CHECKING, Any, Collection, Dict, List, Literal, Mapping, Optional, Set, Type,
+    Union,
 )
-
-from typing_extensions import Literal
 
 import cdedb.validationtypes as vtypes
 from cdedb.common import (
@@ -96,7 +98,7 @@ class GeneralMailinglist:
             **cls.mandatory_validation_fields,
             **cls.optional_validation_fields,
         }.items():
-            if getattr(argtype, "__origin__", None) is list:
+            if typing.get_origin(argtype) is list:
                 ret[field] = "[str]"
             else:
                 ret[field] = "str"
@@ -447,14 +449,11 @@ class EventAssociatedMailinglist(EventAssociatedMeta, EventMailinglist):
         if mailinglist["event_id"] is None:
             return {anid: SubscriptionPolicy.invitation_only for anid in persona_ids}
 
-        ret = {}
-        for persona_id in persona_ids:
-            if bc.event.check_registration_status(
-                    rs, persona_id, mailinglist['event_id'],
-                    mailinglist['registration_stati']):
-                ret[persona_id] = SubscriptionPolicy.subscribable
-            else:
-                ret[persona_id] = SubscriptionPolicy.none
+        ret = bc.event.check_registrations_status(
+            rs, persona_ids, mailinglist['event_id'],
+            mailinglist['registration_stati'])
+        for k, v in ret.items():
+            ret[k] = SubscriptionPolicy.subscribable if v else SubscriptionPolicy.none
 
         return ret
 
