@@ -20,12 +20,22 @@ particular object they could possibly be subscribed to. This determines what act
 they can perform themselves, but also what administrative actions others may perform
 for them.
 """
+# We allow 120 line length here.
+# pylint: disable=line-too-long
 
 import enum
-from gettext import gettext as _
 from typing import Mapping, Optional, Set
 
 from .exceptions import SubscriptionError, SubscriptionInfo
+
+
+def n_(x: str) -> str:
+    """
+    Alias of the identity for i18n.
+    Identity function that shadows the gettext alias to trick pybabel into
+    adding string to the translated strings.
+    """
+    return x
 
 
 @enum.unique
@@ -212,171 +222,171 @@ _ACTION_TARGET_STATE_MAP: Mapping[SubscriptionAction, SubscriptionState] = {
 
 # Errors are identical for all managing actions handling a subscription request.
 _SUBSCRIPTION_REQUEST_ERROR_MAPPING: _StateErrorMapping = {
-    SubscriptionState.subscribed: SubscriptionError(_(
+    SubscriptionState.subscribed: SubscriptionError(n_(
         "subman_managing_not-pending")),
-    SubscriptionState.unsubscribed: SubscriptionError(_(
+    SubscriptionState.unsubscribed: SubscriptionError(n_(
         "subman_managing_not-pending")),
-    SubscriptionState.subscription_override: SubscriptionError(_(
+    SubscriptionState.subscription_override: SubscriptionError(n_(
         "subman_managing_not-pending")),
-    SubscriptionState.unsubscription_override: SubscriptionError(_(
+    SubscriptionState.unsubscription_override: SubscriptionError(n_(
         "subman_managing_not-pending")),
     SubscriptionState.pending: None,
-    SubscriptionState.implicit: SubscriptionError(_(
+    SubscriptionState.implicit: SubscriptionError(n_(
         "subman_managing_not-pending")),
-    SubscriptionState.none: SubscriptionError(_(
+    SubscriptionState.none: SubscriptionError(n_(
         "subman_managing_not-pending")),
 }
 SUBSCRIPTION_ERROR_MATRIX: ActionStateErrorMatrix = {
     SubscriptionAction.add_subscriber: {
-        SubscriptionState.subscribed: SubscriptionInfo(_(
+        SubscriptionState.subscribed: SubscriptionInfo(n_(
             "subman_managing_is-subscribed")),
         SubscriptionState.unsubscribed: None,
-        SubscriptionState.subscription_override: SubscriptionInfo(_(
+        SubscriptionState.subscription_override: SubscriptionInfo(n_(
             "subman_managing_is-subscribed")),
-        SubscriptionState.unsubscription_override: SubscriptionError(_(
+        SubscriptionState.unsubscription_override: SubscriptionError(n_(
             "subman_managing_is-unsubscription-overridden")),
-        SubscriptionState.pending: SubscriptionError(_(
+        SubscriptionState.pending: SubscriptionError(n_(
             "subman_managing_is-pending")),
-        SubscriptionState.implicit: SubscriptionInfo(_(
+        SubscriptionState.implicit: SubscriptionInfo(n_(
             "subman_managing_is-subscribed")),
         SubscriptionState.none: None
     },
     SubscriptionAction.remove_subscriber: {
         SubscriptionState.subscribed: None,
-        SubscriptionState.unsubscribed: SubscriptionInfo(_(
+        SubscriptionState.unsubscribed: SubscriptionInfo(n_(
             "subman_managing_is-unsubscribed")),
-        SubscriptionState.subscription_override: SubscriptionError(_(
+        SubscriptionState.subscription_override: SubscriptionError(n_(
             "subman_managing_is-subscription-overridden")),
-        SubscriptionState.unsubscription_override: SubscriptionInfo(_(
+        SubscriptionState.unsubscription_override: SubscriptionInfo(n_(
             "subman_managing_is-unsubscribed")),
-        SubscriptionState.pending: SubscriptionError(_(
+        SubscriptionState.pending: SubscriptionError(n_(
             "subman_managing_is-pending")),
         SubscriptionState.implicit: None,
-        SubscriptionState.none: SubscriptionInfo(_(
+        SubscriptionState.none: SubscriptionInfo(n_(
             "subman_managing_is-unsubscribed")),
     },
     SubscriptionAction.add_subscription_override: {
         SubscriptionState.subscribed: None,
         SubscriptionState.unsubscribed: None,
-        SubscriptionState.subscription_override: SubscriptionInfo(_(
+        SubscriptionState.subscription_override: SubscriptionInfo(n_(
             "subman_managing_is-subscription-overridden")),
-        SubscriptionState.unsubscription_override: SubscriptionError(_(
+        SubscriptionState.unsubscription_override: SubscriptionError(n_(
             "subman_managing_is-unsubscription-overridden")),
-        SubscriptionState.pending: SubscriptionError(_(
+        SubscriptionState.pending: SubscriptionError(n_(
             "subman_managing_is-pending")),
         SubscriptionState.implicit: None,
         SubscriptionState.none: None,
     },
     SubscriptionAction.remove_subscription_override: {
-        SubscriptionState.subscribed: SubscriptionError(_(
+        SubscriptionState.subscribed: SubscriptionError(n_(
             "subman_managing_not-subscription-overridden")),
-        SubscriptionState.unsubscribed: SubscriptionError(_(
+        SubscriptionState.unsubscribed: SubscriptionError(n_(
             "subman_managing_not-subscription-overridden")),
         SubscriptionState.subscription_override: None,
-        SubscriptionState.unsubscription_override: SubscriptionError(_(
+        SubscriptionState.unsubscription_override: SubscriptionError(n_(
             "subman_managing_not-subscription-overridden")),
-        SubscriptionState.pending: SubscriptionError(_(
+        SubscriptionState.pending: SubscriptionError(n_(
             "subman_managing_not-subscription-overridden")),
-        SubscriptionState.implicit: SubscriptionError(_(
+        SubscriptionState.implicit: SubscriptionError(n_(
             "subman_managing_not-subscription-overridden")),
-        SubscriptionState.none: SubscriptionError(_(
+        SubscriptionState.none: SubscriptionError(n_(
             "subman_managing_not-subscription-overridden")),
     },
     SubscriptionAction.add_unsubscription_override: {
         SubscriptionState.subscribed: None,
         SubscriptionState.unsubscribed: None,
-        SubscriptionState.subscription_override: SubscriptionError(_(
+        SubscriptionState.subscription_override: SubscriptionError(n_(
             "subman_managing_is-subscription-overridden")),
-        SubscriptionState.unsubscription_override: SubscriptionInfo(_(
+        SubscriptionState.unsubscription_override: SubscriptionInfo(n_(
             "subman_managing_is-unsubscription-overridden")),
-        SubscriptionState.pending: SubscriptionError(_(
+        SubscriptionState.pending: SubscriptionError(n_(
             "subman_managing_is-pending")),
         SubscriptionState.implicit: None,
         SubscriptionState.none: None,
     },
     SubscriptionAction.remove_unsubscription_override: {
-        SubscriptionState.subscribed: SubscriptionError(_(
+        SubscriptionState.subscribed: SubscriptionError(n_(
             "subman_managing_not-unsubscription-overridden")),
-        SubscriptionState.unsubscribed: SubscriptionError(_(
+        SubscriptionState.unsubscribed: SubscriptionError(n_(
             "subman_managing_not-unsubscription-overridden")),
-        SubscriptionState.subscription_override: SubscriptionError(_(
+        SubscriptionState.subscription_override: SubscriptionError(n_(
             "subman_managing_not-unsubscription-overridden")),
         SubscriptionState.unsubscription_override: None,
-        SubscriptionState.pending: SubscriptionError(_(
+        SubscriptionState.pending: SubscriptionError(n_(
             "subman_managing_not-unsubscription-overridden")),
-        SubscriptionState.implicit: SubscriptionError(_(
+        SubscriptionState.implicit: SubscriptionError(n_(
             "subman_managing_not-unsubscription-overridden")),
-        SubscriptionState.none: SubscriptionError(_(
+        SubscriptionState.none: SubscriptionError(n_(
             "subman_managing_not-unsubscription-overridden")),
     },
     SubscriptionAction.subscribe: {
-        SubscriptionState.subscribed: SubscriptionInfo(_(
+        SubscriptionState.subscribed: SubscriptionInfo(n_(
             "subman_self_is-subscribed")),
         SubscriptionState.unsubscribed: None,
-        SubscriptionState.subscription_override: SubscriptionInfo(_(
+        SubscriptionState.subscription_override: SubscriptionInfo(n_(
             "subman_self_is-subscribed")),
-        SubscriptionState.unsubscription_override: SubscriptionError(_(
+        SubscriptionState.unsubscription_override: SubscriptionError(n_(
             "subman_self_is-unsubscription-overridden")),
         SubscriptionState.pending: None,
-        SubscriptionState.implicit: SubscriptionInfo(_(
+        SubscriptionState.implicit: SubscriptionInfo(n_(
             "subman_self_is-subscribed")),
         SubscriptionState.none: None,
     },
     SubscriptionAction.request_subscription: {
-        SubscriptionState.subscribed: SubscriptionInfo(_(
+        SubscriptionState.subscribed: SubscriptionInfo(n_(
             "subman_self_is-subscribed")),
         SubscriptionState.unsubscribed: None,
-        SubscriptionState.subscription_override: SubscriptionInfo(_(
+        SubscriptionState.subscription_override: SubscriptionInfo(n_(
             "subman_self_is-subscribed")),
-        SubscriptionState.unsubscription_override: SubscriptionError(_(
+        SubscriptionState.unsubscription_override: SubscriptionError(n_(
             "subman_self_is-unsubscription-overridden")),
-        SubscriptionState.pending: SubscriptionInfo(_(
+        SubscriptionState.pending: SubscriptionInfo(n_(
             "subman_self_is-pending")),
-        SubscriptionState.implicit: SubscriptionInfo(_(
+        SubscriptionState.implicit: SubscriptionInfo(n_(
             "subman_self_is-subscribed")),
         SubscriptionState.none: None,
     },
     SubscriptionAction.unsubscribe: {
         SubscriptionState.subscribed: None,
-        SubscriptionState.unsubscribed: SubscriptionInfo(_(
+        SubscriptionState.unsubscribed: SubscriptionInfo(n_(
             "subman_self_is-unsubscribed")),
         # subscription_override should only block you from being unsubscribed
         # automatically. A user is still able to unsubscribe manually.
         # (Unless the list is mandatory).
         SubscriptionState.subscription_override: None,
-        SubscriptionState.unsubscription_override: SubscriptionInfo(_(
+        SubscriptionState.unsubscription_override: SubscriptionInfo(n_(
             "subman_self_is-unsubscribed")),
-        SubscriptionState.pending: SubscriptionInfo(_(
+        SubscriptionState.pending: SubscriptionInfo(n_(
             "subman_self_is-unsubscribed")),
         SubscriptionState.implicit: None,
-        SubscriptionState.none: SubscriptionInfo(_(
+        SubscriptionState.none: SubscriptionInfo(n_(
             "subman_self_is-unsubscribed")),
     },
     SubscriptionAction.reset: {
         SubscriptionState.subscribed: None,
         SubscriptionState.unsubscribed: None,
-        SubscriptionState.subscription_override: SubscriptionError(_(
+        SubscriptionState.subscription_override: SubscriptionError(n_(
             "subman_managing_is-subscription-overridden")),
-        SubscriptionState.unsubscription_override: SubscriptionError(_(
+        SubscriptionState.unsubscription_override: SubscriptionError(n_(
             "subman_managing_is-unsubscription-overridden")),
-        SubscriptionState.pending: SubscriptionError(_(
+        SubscriptionState.pending: SubscriptionError(n_(
             "subman_managing_is-pending")),
         SubscriptionState.implicit: None,
         SubscriptionState.none: None,
     },
     SubscriptionAction.cancel_request: {
-        SubscriptionState.subscribed: SubscriptionError(_(
+        SubscriptionState.subscribed: SubscriptionError(n_(
             "subman_self_not-pending")),
-        SubscriptionState.unsubscribed: SubscriptionError(_(
+        SubscriptionState.unsubscribed: SubscriptionError(n_(
             "subman_self_not-pending")),
-        SubscriptionState.subscription_override: SubscriptionError(_(
+        SubscriptionState.subscription_override: SubscriptionError(n_(
             "subman_self_not-pending")),
-        SubscriptionState.unsubscription_override: SubscriptionError(_(
+        SubscriptionState.unsubscription_override: SubscriptionError(n_(
             "subman_self_not-pending")),
         SubscriptionState.pending: None,
-        SubscriptionState.implicit: SubscriptionError(_(
+        SubscriptionState.implicit: SubscriptionError(n_(
             "subman_self_not-pending")),
-        SubscriptionState.none: SubscriptionError(_(
+        SubscriptionState.none: SubscriptionError(n_(
             "subman_self_not-pending")),
 },
     SubscriptionAction.approve_request: _SUBSCRIPTION_REQUEST_ERROR_MAPPING,
