@@ -1,3 +1,5 @@
+"""data types for the CdEDB project"""
+
 import datetime
 import decimal
 from typing import Any, Dict, List, NewType as _NewType, Type
@@ -66,11 +68,14 @@ ByFieldDatatype = _NewType("ByFieldDatatype", str)
 
 Persona = _NewType("Persona", CdEDBObject)
 GenesisCase = _NewType("GenesisCase", CdEDBObject)
+BatchAdmissionEntry = _NewType("BatchAdmissionEntry", CdEDBObject)
 PrivilegeChange = _NewType("PrivilegeChange", CdEDBObject)
 Period = _NewType("Period", CdEDBObject)
 ExPuls = _NewType("ExPuls", CdEDBObject)
+MoneyTransferEntry = _NewType("MoneyTransferEntry", CdEDBObject)
 Lastschrift = _NewType("Lastschrift", CdEDBObject)
 LastschriftTransaction = _NewType("LastschriftTransaction", CdEDBObject)
+LastschriftTransactionEntry = _NewType("LastschriftTransactionEntry", CdEDBObject)
 SepaTransactions = _NewType("SepaTransactions", List[CdEDBObject])
 SepaMeta = _NewType("SepaMeta", CdEDBObject)
 MetaInfo = _NewType("MetaInfo", CdEDBObject)
@@ -87,15 +92,21 @@ Registration = _NewType("Registration", CdEDBObject)
 RegistrationPart = _NewType("RegistrationPart", CdEDBObject)
 RegistrationTrack = _NewType("RegistrationTrack", CdEDBObject)
 EventAssociatedFields = _NewType("EventAssociatedFields", CdEDBObject)
+FeeBookingEntry = _NewType("FeeBookingEntry", CdEDBObject)
 LodgementGroup = _NewType("LodgementGroup", CdEDBObject)
 Lodgement = _NewType("Lodgement", CdEDBObject)
-Questionnaire = _NewType("Questionnaire", Dict[int, Any])  # TODO maybe cast keys to str
+QuestionnaireRow = _NewType("QuestionnaireRow", CdEDBObject)
+# TODO maybe cast keys to str
+Questionnaire = _NewType("Questionnaire", Dict[int, List[QuestionnaireRow]])
 
 SerializedEvent = _NewType("SerializedEvent", CdEDBObject)
 SerializedEventUpload = _NewType("SerializedEventUpload", SerializedEvent)
 SerializedPartialEvent = _NewType("SerializedPartialEvent", CdEDBObject)
 SerializedPartialEventUpload = _NewType(
     "SerializedPartialEventUpload", SerializedPartialEvent)
+SerializedEventQuestionnaire = _NewType("SerializedEventQuestionnaire", CdEDBObject)
+SerializedEventQuestionnaireUpload = _NewType(
+    "SerializedEventQuestionnaireUpload", SerializedEventQuestionnaire)
 
 PartialCourse = _NewType("PartialCourse", CdEDBObject)
 PartialLodgementGroup = _NewType("PartialLodgementGroup", CdEDBObject)
