@@ -1,4 +1,5 @@
 #!/usr/bin/env python3
+"""General testing utilities for CdEDB2 testsuite"""
 
 import collections.abc
 import copy
@@ -19,7 +20,6 @@ import shutil
 import subprocess
 import sys
 import tempfile
-import time
 import unittest
 import urllib.parse
 from typing import (
@@ -47,7 +47,7 @@ from cdedb.config import BasicConfig, Config, SecretsConfig
 from cdedb.database import DATABASE_ROLES
 from cdedb.database.connection import connection_pool_factory
 from cdedb.frontend.application import Application
-from cdedb.frontend.common import AbstractFrontend
+from cdedb.frontend.common import AbstractFrontend, Worker, setup_translations
 from cdedb.frontend.cron import CronFrontend
 from cdedb.query import QueryOperators
 from cdedb.script import setup
@@ -58,6 +58,7 @@ _SECRETSCONF = SecretsConfig()
 # TODO: use TypedDict to specify UserObject.
 UserObject = Mapping[str, Any]
 UserIdentifier = Union[UserObject, str, int]
+LinkIdentifier = Union[MutableMapping[str, Any], str]
 
 # This is to be used in place of `self.key` for anonymous requests. It makes mypy happy.
 ANONYMOUS = cast(RequestState, None)
@@ -139,15 +140,14 @@ def _make_backend_shim(backend: B, internal: bool = False) -> B:
     This is similar to the normal make_proxy but encorporates a different
     wrapper.
     """
+    # pylint: disable=protected-access
 
     sessionproxy = SessionBackend(backend.conf._configpath)
     secrets = SecretsConfig(backend.conf._configpath)
     connpool = connection_pool_factory(
         backend.conf["CDB_DATABASE_NAME"], DATABASE_ROLES,
         secrets, backend.conf["DB_PORT"])
-    translator = gettext.translation(
-        'cdedb', languages=['de'],
-        localedir=str(backend.conf["REPOSITORY_PATH"] / 'i18n'))
+    translations = setup_translations(backend.conf)
 
     def setup_requeststate(key: Optional[str], ip: str = "127.0.0.0"
                            ) -> RequestState:
@@ -171,11 +171,19 @@ def _make_backend_shim(backend: B, internal: bool = False) -> B:
             apitoken = key
 
         rs = RequestState(
-            sessionkey=sessionkey, apitoken=apitoken, user=user,
-            request=None, notifications=[], mapadapter=None,  # type: ignore
-            requestargs=None, errors=[], values=None, lang="de",
-            gettext=translator.gettext, ngettext=translator.ngettext,
-            coders=None, begin=now())
+            sessionkey=sessionkey,
+            apitoken=apitoken,
+            user=user,
+            request=None,  # type: ignore[arg-type]
+            notifications=[],
+            mapadapter=None,  # type: ignore[arg-type]
+            requestargs=None,
+            errors=[],
+            values=None,
+            begin=now(),
+            lang="de",
+            translations=translations,
+        )
         rs._conn = connpool[roles_to_db_role(rs.user.roles)]
         rs.conn = rs._conn
         if "event" in rs.user.roles and hasattr(backend, "orga_info"):
@@ -209,7 +217,11 @@ def _make_backend_shim(backend: B, internal: bool = False) -> B:
             @functools.wraps(attr)
             def wrapper(key: Optional[str], *args: Any, **kwargs: Any) -> Any:
                 rs = setup_requeststate(key)
-                return attr(rs, *args, **kwargs)
+                try:
+                    return attr(rs, *args, **kwargs)
+                except FileNotFoundError as e:
+                    raise RuntimeError("Did you forget to add a `@storage` decorator to"
+                                       " the test?") from e
 
             return wrapper
 
@@ -265,23 +277,22 @@ class BasicTest(unittest.TestCase):
 
         # Turn Iterator into Collection and ensure consistent order.
         keys = tuple(keys)
+        if not keys:
+            keys = tuple(next(iter(_SAMPLE_DATA[table].values())).keys())
         ret = {}
         for anid in ids:
-            if keys:
-                r = {}
-                for k in keys:
-                    r[k] = copy.deepcopy(_SAMPLE_DATA[table][anid][k])
-                    if table == 'core.personas':
-                        if k == 'balance':
-                            r[k] = decimal.Decimal(r[k])
-                        if k == 'birthday':
-                            r[k] = datetime.date.fromisoformat(r[k])
-                    if k in {'ctime', 'atime', 'vote_begin', 'vote_end',
-                             'vote_extension_end'}:
-                        r[k] = parse_datetime(r[k])
-                ret[anid] = r
-            else:
-                ret[anid] = copy.deepcopy(_SAMPLE_DATA[table][anid])
+            r = {}
+            for k in keys:
+                r[k] = copy.deepcopy(_SAMPLE_DATA[table][anid][k])
+                if table == 'core.personas':
+                    if k == 'balance':
+                        r[k] = decimal.Decimal(r[k])
+                    if k == 'birthday':
+                        r[k] = datetime.date.fromisoformat(r[k])
+                if k in {'ctime', 'atime', 'vote_begin', 'vote_end',
+                         'vote_extension_end', 'signup_end'}:
+                    r[k] = parse_datetime(r[k])
+            ret[anid] = r
         return ret
 
     def get_sample_datum(self, table: str, id_: int) -> CdEDBObject:
@@ -323,6 +334,7 @@ class BackendTest(CdEDBTest):
     pastevent: ClassVar[PastEventBackend]
     ml: ClassVar[MlBackend]
     assembly: ClassVar[AssemblyBackend]
+    translations: ClassVar[Mapping[str, gettext.NullTranslations]]
     user: UserObject
     key: RequestState
 
@@ -339,6 +351,7 @@ class BackendTest(CdEDBTest):
         # Workaround to make orga info available for calls into the MLBackend.
         cls.ml.orga_info = lambda rs, persona_id: cls.event.orga_info(  # type: ignore
             rs.sessionkey, persona_id)
+        cls.translations = setup_translations(cls.conf)
 
     def setUp(self) -> None:
         """Reset login state."""
@@ -369,6 +382,23 @@ class BackendTest(CdEDBTest):
         users = {get_user(i)["id"] for i in identifiers}
         return self.user.get("id", -1) in users
 
+    def assertLogEqual(self, log_expectation: Sequence[CdEDBObject], realm: str,
+                       **kwargs: Any) -> None:
+        """Helper to compare a log expectation to the actual thing."""
+        _, log = getattr(self, realm).retrieve_log(self.key, **kwargs)
+        log_expectation = tuple(log_expectation)
+        for e in log:
+            del e['id']
+        for e in log_expectation:
+            if 'ctime' not in e:
+                e['ctime'] = nearly_now()
+            if 'submitted_by' not in e:
+                e['submitted_by'] = self.user['id']
+            for k in ('persona_id', 'change_note'):
+                if k not in e:
+                    e[k] = None
+        self.assertEqual(log_expectation, log)
+
     @staticmethod
     def initialize_raw_backend(backendcls: Type[SessionBackend]
                                ) -> SessionBackend:
@@ -391,6 +421,7 @@ USER_DICT: Dict[str, UserObject] = {
         'display_name': "Anton",
         'given_names': "Anton Armin A.",
         'family_name': "Administrator",
+        'default_name_format': "Anton Administrator",
     },
     "berta": {
         'id': 2,
@@ -400,6 +431,7 @@ USER_DICT: Dict[str, UserObject] = {
         'display_name': "Bertå",
         'given_names': "Bertålotta",
         'family_name': "Beispiel",
+        'default_name_format': "Bertå Beispiel",
     },
     "charly": {
         'id': 3,
@@ -409,6 +441,7 @@ USER_DICT: Dict[str, UserObject] = {
         'display_name': "Charly",
         'given_names': "Charly C.",
         'family_name': "Clown",
+        'default_name_format': "Charly Clown",
     },
     "daniel": {
         'id': 4,
@@ -418,15 +451,17 @@ USER_DICT: Dict[str, UserObject] = {
         'display_name': "Daniel",
         'given_names': "Daniel D.",
         'family_name': "Dino",
+        'default_name_format': "Daniel Dino",
     },
     "emilia": {
         'id': 5,
         'DB-ID': "DB-5-1",
         'username': "emilia@example.cde",
         'password': "secret",
-        'display_name': "Emilia",
+        'display_name': "Emmy",
         'given_names': "Emilia E.",
         'family_name': "Eventis",
+        'default_name_format': "Emilia E. Eventis",
     },
     "ferdinand": {
         'id': 6,
@@ -436,6 +471,7 @@ USER_DICT: Dict[str, UserObject] = {
         'display_name': "Ferdinand",
         'given_names': "Ferdinand F.",
         'family_name': "Findus",
+        'default_name_format': "Ferdinand Findus",
     },
     "garcia": {
         'id': 7,
@@ -445,6 +481,7 @@ USER_DICT: Dict[str, UserObject] = {
         'display_name': "Garcia",
         'given_names': "Garcia G.",
         'family_name': "Generalis",
+        'default_name_format': "Garcia Generalis",
     },
     "hades": {
         'id': 8,
@@ -454,6 +491,7 @@ USER_DICT: Dict[str, UserObject] = {
         'display_name': None,
         'given_names': "Hades",
         'family_name': "Hell",
+        'default_name_format': "Hades Hell",
     },
     "inga": {
         'id': 9,
@@ -463,6 +501,7 @@ USER_DICT: Dict[str, UserObject] = {
         'display_name': "Inga",
         'given_names': "Inga",
         'family_name': "Iota",
+        'default_name_format': "Inga Iota",
     },
     "janis": {
         'id': 10,
@@ -472,6 +511,7 @@ USER_DICT: Dict[str, UserObject] = {
         'display_name': "Janis",
         'given_names': "Janis",
         'family_name': "Jalapeño",
+        'default_name_format': "Janis Jalapeño",
     },
     "kalif": {
         'id': 11,
@@ -481,6 +521,7 @@ USER_DICT: Dict[str, UserObject] = {
         'display_name': "Kalif",
         'given_names': "Kalif ibn al-Ḥasan",
         'family_name': "Karabatschi",
+        'default_name_format': "Kalif Karabatschi",
     },
     "lisa": {
         'id': 12,
@@ -490,6 +531,7 @@ USER_DICT: Dict[str, UserObject] = {
         'display_name': "Lisa",
         'given_names': "Lisa",
         'family_name': "Lost",
+        'default_name_format': "Lisa Lost",
     },
     "martin": {
         'id': 13,
@@ -499,6 +541,7 @@ USER_DICT: Dict[str, UserObject] = {
         'display_name': "Martin",
         'given_names': "Martin",
         'family_name': "Meister",
+        'default_name_format': "Martin Meister",
     },
     "nina": {
         'id': 14,
@@ -508,6 +551,7 @@ USER_DICT: Dict[str, UserObject] = {
         'display_name': "Nina",
         'given_names': "Nina",
         'family_name': "Neubauer",
+        'default_name_format': "Nina Neubauer",
     },
     "olaf": {
         'id': 15,
@@ -517,6 +561,7 @@ USER_DICT: Dict[str, UserObject] = {
         'display_name': "Olaf",
         'given_names': "Olaf",
         'family_name': "Olafson",
+        'default_name_format': "Olaf Olafson",
     },
     "paul": {
         'id': 16,
@@ -526,6 +571,7 @@ USER_DICT: Dict[str, UserObject] = {
         'display_name': "Paul",
         'given_names': "Paulchen",
         'family_name': "Panther",
+        'default_name_format': "Paul Panther",
     },
     "quintus": {
         'id': 17,
@@ -535,6 +581,7 @@ USER_DICT: Dict[str, UserObject] = {
         'display_name': "Quintus",
         'given_names': "Quintus",
         'family_name': "da Quirm",
+        'default_name_format': "Quintus da Quirm",
     },
     "rowena": {
         'id': 18,
@@ -544,6 +591,7 @@ USER_DICT: Dict[str, UserObject] = {
         'display_name': "Rowena",
         'given_names': "Rowena",
         'family_name': "Ravenclaw",
+        'default_name_format': "Rowena Ravenclaw",
     },
     "vera": {
         'id': 22,
@@ -553,6 +601,7 @@ USER_DICT: Dict[str, UserObject] = {
         'display_name': "Vera",
         'given_names': "Vera",
         'family_name': "Verwaltung",
+        'default_name_format': "Vera Verwaltung",
     },
     "werner": {
         'id': 23,
@@ -562,6 +611,7 @@ USER_DICT: Dict[str, UserObject] = {
         'display_name': "Werner",
         'given_names': "Werner",
         'family_name': "Wahlleitung",
+        'default_name_format': "Werner Wahlleitung",
     },
     "annika": {
         'id': 27,
@@ -571,6 +621,7 @@ USER_DICT: Dict[str, UserObject] = {
         'display_name': "Annika",
         'given_names': "Annika",
         'family_name': "Akademieteam",
+        'default_name_format': "Annika Akademieteam",
     },
     "farin": {
         'id': 32,
@@ -580,6 +631,7 @@ USER_DICT: Dict[str, UserObject] = {
         'display_name': "Farin",
         'given_names': "Farin",
         'family_name': "Finanzvorstand",
+        'default_name_format': "Farin Finanzvorstand",
     },
     "viktor": {
         'id': 48,
@@ -589,6 +641,7 @@ USER_DICT: Dict[str, UserObject] = {
         'display_name': "Viktor",
         'given_names': "Viktor",
         'family_name': "Versammlungsadmin",
+        'default_name_format': "Viktor Versammlungsadmin",
     },
     "akira": {
         'id': 100,
@@ -598,6 +651,7 @@ USER_DICT: Dict[str, UserObject] = {
         'display_name': "Akira",
         'given_names': "Akira",
         'family_name': "Abukara",
+        'default_name_format': "Akira Abukara",
     },
     "anonymous": {
         'id': None,
@@ -672,8 +726,7 @@ def execsql(sql: AnyStr) -> None:
     """Execute arbitrary SQL-code on the test database."""
     psql = ("/cdedb2/bin/execute_sql_script.py",
             "--username", "cdb", "--dbname", os.environ['CDEDB_TEST_DATABASE'])
-    # TODO: remove the type: ignore in a newer mypy version (0.800 or higher)
-    mode = 'wb' if isinstance(sql, bytes) else 'w'  # type: ignore[unreachable]
+    mode = 'wb' if isinstance(sql, bytes) else 'w'
     with tempfile.NamedTemporaryFile(mode=mode, suffix='.sql') as sql_file:
         sql_file.write(sql)
         sql_file.flush()
@@ -699,6 +752,7 @@ class FrontendTest(BackendTest):
     response: webtest.TestResponse
     app_extra_environ = {
         'REMOTE_ADDR': "127.0.0.0",
+        'HTTP_HOST': "localhost",
         'SERVER_PROTOCOL': "HTTP/1.1",
         'wsgi.url_scheme': 'https'}
 
@@ -750,7 +804,8 @@ class FrontendTest(BackendTest):
             # path without host but with query string - capped at 64 chars
             # To enhance readability, we mark most chars as safe. All special chars are
             # allowed in linux file paths, but sadly windows is more restrictive...
-            url = urllib.parse.quote(self.response.request.path_qs, safe='/;@&=+$,~')[:64]
+            url = urllib.parse.quote(
+                self.response.request.path_qs, safe='/;@&=+$,~')[:64]
             # since / chars are forbidden in file paths, we replace them by _
             url = url.replace('/', '_')
             # create a temporary file in scrap_path with url as a prefix
@@ -782,6 +837,16 @@ class FrontendTest(BackendTest):
         self.response = self.response.maybe_follow(**kwargs)
         if self.response != oldresponse:
             self._log_generation_time(oldresponse)
+
+    def assertRedirect(self, url: str, *args: Any, target_url: str,
+                       verbose: bool = False, **kwargs: Any) -> webtest.TestResponse:
+        """Checck that a GET-request to the url returns a redirect to the target url."""
+        response: webtest.TestResponse = self.app.get(url, *args, **kwargs)
+        self.assertLessEqual(300, response.status_int)
+        self.assertGreater(400, response.status_int)
+        self.assertIn("You should be redirected", response)
+        self.assertIn(target_url, response)
+        return response
 
     def post(self, url: str, *args: Any, verbose: bool = False, **kwargs: Any) -> None:
         """Directly send a POST-request.
@@ -829,8 +894,7 @@ class FrontendTest(BackendTest):
                 raise AssertionError(
                     "Post request did not produce success notification.")
 
-    def traverse(self, *links: Union[MutableMapping[str, Any], str],
-                 verbose: bool = False) -> None:
+    def traverse(self, *links: LinkIdentifier, verbose: bool = False) -> None:
         """Follow a sequence of links, described by their kwargs.
 
         A link can also be just a string, in which case that string is assumed
@@ -907,8 +971,7 @@ class FrontendTest(BackendTest):
         f['phrase'] = u["DB-ID"]
         self.submit(f)
         if check:
-            self.assertTitle("{} {}".format(u['given_names'],
-                                            u['family_name']))
+            self.assertTitle(u['default_name_format'])
 
     def realm_admin_view_profile(self, user: str, realm: str,
                                  check: bool = True, verbose: bool = False
@@ -933,8 +996,7 @@ class FrontendTest(BackendTest):
         self.submit(f, verbose=verbose)
         self.traverse({'description': 'Profil'}, verbose=verbose)
         if check:
-            self.assertTitle("{} {}".format(u['given_names'],
-                                            u['family_name']))
+            self.assertTitle(u['default_name_format'])
 
     def _fetch_mail(self) -> List[email.message.EmailMessage]:
         """
@@ -1241,7 +1303,7 @@ class FrontendTest(BackendTest):
         f = self.response.forms['logshowform']
         # use internal value property as I don't see a way to get the
         # checkbox value otherwise
-        codes = [field._value for field in f.fields['codes']]
+        codes = [field._value for field in f.fields['codes']]  # pylint: disable=protected-access
         f['codes'] = codes
         self.assertGreater(len(codes), 1)
         self.submit(f)
@@ -1392,23 +1454,24 @@ class FrontendTest(BackendTest):
                     value=button['value'])
         return button
 
-    def reload_and_check_form(self, form: webtest.Form, link: Union[CdEDBObject, str],
-                              max_tries: int = 42, waittime: float = 0.1,
-                              fail: bool = True) -> None:
-        """Helper to repeatedly reload a page until a certain form is present.
+    def join_worker_thread(self, worker_name: str, link: LinkIdentifier, *,
+                           realm: str = "cde", timeout: float = 2) -> None:
+        """Wait for the specified Worker thread to finish.
 
-        This is mostly required for the "Semesterverwaltung".
+        :param realm: specify to which realm the Worker belongs. Currently only the
+            CdEFrontend uses Workers.
+        :param timeout: pecificy a maximum wait time for the thread to end. In our
+            testing environment Worker threads should not take longer than a couple
+            seconds.
         """
-        count = 0
-        while count < max_tries:
-            time.sleep(waittime)
-            self.traverse(link)
-            if form in self.response.forms:
-                break
-            count += 1
-        else:
-            if fail:
-                self.fail(f"Form {form} not found after {count} reloads.")
+        ref = Worker.active_workers[worker_name]
+        worker = ref()
+        if worker:
+            worker.join(timeout)
+            if worker.is_alive():
+                self.fail(f"Worker {realm}/{worker_name} still active after {timeout}"
+                          f" seconds.")
+        self.traverse(link)
 
 
 class MultiAppFrontendTest(FrontendTest):
