@@ -1,4 +1,5 @@
 #!/usr/bin/env python3
+# pylint: disable=protected-access,missing-module-docstring
 
 from typing import Collection, Set, cast
 
@@ -43,7 +44,8 @@ class TestMlBackend(BackendTest):
             self.ml.merge_accounts(self.key,
                                    source_persona_id=USER_DICT['nina']['id'],
                                    target_persona_id=berta_id)
-        self.assertEqual("Source User is admin and can not be merged.", str(e.exception))
+        self.assertEqual("Source User is admin and can not be merged.",
+                         str(e.exception))
 
         with self.assertRaises(ValueError) as e:
             self.ml.merge_accounts(self.key,
@@ -340,7 +342,7 @@ class TestMlBackend(BackendTest):
             'title': 'Proletarier aller Länder',
         }
         self.assertLess(0, self.ml.create_mailinglist(self.key, new_data))
-        assert isinstance(new_data['moderators'], Set)
+        assert isinstance(new_data['moderators'], Set)  # pylint: disable=isinstance-second-argument-not-valid-type
         new_data['moderators'] |= {100000}
         with self.assertRaises(ValueError):
             self.ml.create_mailinglist(self.key, new_data)
@@ -387,7 +389,7 @@ class TestMlBackend(BackendTest):
         for ml_id in ml_ids:
             with self.subTest(ml_id=ml_id):
                 expectation = self.ml.get_subscription_states(self.key, ml_id)
-                self.ml.write_subscription_states(self.key, ml_id)
+                self.ml.write_subscription_states(self.key, (ml_id,))
                 result = self.ml.get_subscription_states(self.key, ml_id)
 
                 self.assertEqual(expectation, result)
@@ -626,7 +628,8 @@ class TestMlBackend(BackendTest):
                          state=SS.unsubscribed, kind="info")
 
         # This does some basic override testing.
-        self._change_sub(self.user['id'], mailinglist_id, SA.add_unsubscription_override,
+        self._change_sub(self.user['id'], mailinglist_id,
+                         SA.add_unsubscription_override,
                          state=SS.unsubscription_override)
         self._change_sub(self.user['id'], mailinglist_id, SA.subscribe,
                          state=SS.unsubscription_override, kind="error")
@@ -639,8 +642,8 @@ class TestMlBackend(BackendTest):
         # transition between non-core states are forbidden
         self._change_sub(self.user['id'], mailinglist_id, SA.add_subscription_override,
                          state=SS.unsubscription_override, kind="error")
-        self._change_sub(self.user['id'], mailinglist_id, SA.remove_unsubscription_override,
-                         state=SS.unsubscribed)
+        self._change_sub(self.user['id'], mailinglist_id,
+                         SA.remove_unsubscription_override, state=SS.unsubscribed)
         self._change_sub(self.user['id'], mailinglist_id, SA.add_subscription_override,
                          state=SS.subscription_override)
         self._change_sub(self.user['id'], mailinglist_id, SA.remove_subscriber,
@@ -651,7 +654,8 @@ class TestMlBackend(BackendTest):
                          state=SS.subscription_override, kind="info")
         self._change_sub(self.user['id'], mailinglist_id, SA.unsubscribe,
                          state=SS.unsubscribed)
-        self._change_sub(self.user['id'], mailinglist_id, SA.remove_subscription_override,
+        self._change_sub(self.user['id'], mailinglist_id,
+                         SA.remove_subscription_override,
                          state=SS.unsubscribed, kind="error")
 
         # You cannot request subscriptions to such lists
@@ -670,21 +674,25 @@ class TestMlBackend(BackendTest):
                          state=SS.unsubscribed, kind="info")
 
         # This does more override management testing
-        self._change_sub(self.user['id'], mailinglist_id, SA.remove_unsubscription_override,
+        self._change_sub(self.user['id'], mailinglist_id,
+                         SA.remove_unsubscription_override,
                          state=SS.unsubscribed, kind="error")
-        self._change_sub(self.user['id'], mailinglist_id, SA.remove_subscription_override,
+        self._change_sub(self.user['id'], mailinglist_id,
+                         SA.remove_subscription_override,
                          state=SS.unsubscribed, kind="error")
         self._change_sub(self.user['id'], mailinglist_id, SA.subscribe,
                          state=SS.subscribed)
-        self._change_sub(self.user['id'], mailinglist_id, SA.add_unsubscription_override,
+        self._change_sub(self.user['id'], mailinglist_id,
+                         SA.add_unsubscription_override,
                          state=SS.unsubscription_override)
-        self._change_sub(self.user['id'], mailinglist_id, SA.remove_unsubscription_override,
-                         state=SS.unsubscribed)
+        self._change_sub(self.user['id'], mailinglist_id,
+                         SA.remove_unsubscription_override, state=SS.unsubscribed)
         self._change_sub(self.user['id'], mailinglist_id, SA.add_subscription_override,
                          state=SS.subscription_override)
-        self._change_sub(self.user['id'], mailinglist_id, SA.remove_subscription_override,
-                         state=SS.subscribed)
-        self._change_sub(self.user['id'], mailinglist_id, SA.remove_subscription_override,
+        self._change_sub(self.user['id'], mailinglist_id,
+                         SA.remove_subscription_override, state=SS.subscribed)
+        self._change_sub(self.user['id'], mailinglist_id,
+                         SA.remove_subscription_override,
                          state=SS.subscribed, kind="error")
 
         # This tests the unsubscription reset.
@@ -814,14 +822,15 @@ class TestMlBackend(BackendTest):
                          state=SS.unsubscribed)
 
         # Test blocks
-        self._change_sub(self.user['id'], mailinglist_id, SA.add_unsubscription_override,
+        self._change_sub(self.user['id'], mailinglist_id,
+                         SA.add_unsubscription_override,
                          state=SS.unsubscription_override)
         self._change_sub(self.user['id'], mailinglist_id, SA.subscribe,
                          state=SS.unsubscription_override, kind="error")
         self._change_sub(self.user['id'], mailinglist_id, SA.add_subscriber,
                          state=SS.unsubscription_override, kind="error")
-        self._change_sub(self.user['id'], mailinglist_id, SA.remove_unsubscription_override,
-                         state=SS.unsubscribed)
+        self._change_sub(self.user['id'], mailinglist_id,
+                         SA.remove_unsubscription_override, state=SS.unsubscribed)
 
         # Test forced subscriptions
         self._change_sub(self.user['id'], mailinglist_id, SA.subscribe,
@@ -841,7 +850,7 @@ class TestMlBackend(BackendTest):
 
         for persona_id in {17, 27, 32}:
             self._change_sub(persona_id, mailinglist_id, SA.reset, state=SS.none)
-        self.ml.write_subscription_states(self.key, mailinglist_id)
+        self.ml.write_subscription_states(self.key, (mailinglist_id,))
         for persona_id in {17, 27, 32}:
             new_state = self.ml.get_subscription(self.key, persona_id=persona_id,
                                                  mailinglist_id=mailinglist_id)
@@ -852,7 +861,7 @@ class TestMlBackend(BackendTest):
                              state=SS.subscription_override)
             self._change_sub(persona_id, mailinglist_id,
                              SA.remove_subscription_override, state=SS.subscribed)
-        self.ml.write_subscription_states(self.key, mailinglist_id)
+        self.ml.write_subscription_states(self.key, (mailinglist_id,))
         for persona_id in {17, 27, 32}:
             new_state = self.ml.get_subscription(self.key, persona_id=persona_id,
                                                  mailinglist_id=mailinglist_id)
@@ -894,7 +903,7 @@ class TestMlBackend(BackendTest):
 
         # For admins, some shallow cron testing
         if not self.user_in(2):
-            self.ml.write_subscription_states(self.key, mailinglist_id)
+            self.ml.write_subscription_states(self.key, (mailinglist_id,))
             self._check_state(self.user['id'], mailinglist_id, SS.subscribed)
 
     @as_users('nina')
@@ -915,7 +924,7 @@ class TestMlBackend(BackendTest):
                          state=SS.subscription_override)
 
         # Cron testing
-        self.ml.write_subscription_states(self.key, mailinglist_id)
+        self.ml.write_subscription_states(self.key, (mailinglist_id,))
         self._check_state(self.user['id'], mailinglist_id, SS.subscription_override)
 
         # It is impossible to unsubscribe normally
@@ -925,11 +934,11 @@ class TestMlBackend(BackendTest):
                          state=SS.subscription_override, kind="error")
 
         # Remove subscription
-        self._change_sub(self.user['id'], mailinglist_id, SA.remove_subscription_override,
-                         state=SS.subscribed)
+        self._change_sub(self.user['id'], mailinglist_id,
+                         SA.remove_subscription_override, state=SS.subscribed)
 
         # Cron testing
-        self.ml.write_subscription_states(self.key, mailinglist_id)
+        self.ml.write_subscription_states(self.key, (mailinglist_id,))
         self._check_state(self.user['id'], mailinglist_id, SS.none)
 
     @as_users("anton")
@@ -960,7 +969,7 @@ class TestMlBackend(BackendTest):
         self._change_sub(self.user['id'], ml_id, SA.subscribe,
                          state=SS.subscribed, kind="info")
 
-        self.ml.write_subscription_states(self.key, ml_id)
+        self.ml.write_subscription_states(self.key, (ml_id,))
 
         expectation = {
             1: SS.subscribed,
@@ -1043,10 +1052,10 @@ class TestMlBackend(BackendTest):
         admin_key = cast(RequestState, self.login("viktor"))
         # Remove all assembly presiders
         self.assembly.remove_assembly_presider(admin_key, assembly_id, 23)
-        self.ml.write_subscription_states(self.key, ml_id)
+        self.ml.write_subscription_states(self.key, (ml_id,))
         self.assertEqual({}, self.ml.get_subscription_states(self.key, ml_id))
         self.assembly.add_assembly_presiders(admin_key, assembly_id, {1})
-        self.ml.write_subscription_states(self.key, ml_id)
+        self.ml.write_subscription_states(self.key, (ml_id,))
 
         expectation = {
             1: SS.implicit,
@@ -1183,7 +1192,7 @@ class TestMlBackend(BackendTest):
         self.assertEqual(result, expectation)
 
         self.assertLess(
-            0, self.ml.write_subscription_states(self.key, mailinglist_id))
+            0, self.ml.write_subscription_states(self.key, (mailinglist_id,)))
 
         expectation = {
             1: SS.subscribed,
@@ -1226,7 +1235,7 @@ class TestMlBackend(BackendTest):
         self.assertEqual(result, expectation)
 
         self.assertLess(
-            0, self.ml.write_subscription_states(self.key, mailinglist_id))
+            0, self.ml.write_subscription_states(self.key, (mailinglist_id,)))
 
         expectation = {
             1: SS.implicit,
@@ -1256,7 +1265,7 @@ class TestMlBackend(BackendTest):
         self.assertEqual(result, expectation)
 
         self.assertLess(
-            0, self.ml.write_subscription_states(self.key, mailinglist_id))
+            0, self.ml.write_subscription_states(self.key, (mailinglist_id,)))
 
         expectation = {
             1: SS.implicit,
@@ -1281,7 +1290,7 @@ class TestMlBackend(BackendTest):
             for persona_id in removes
         ]
         self.ml._set_subscriptions(self.key, data)
-        self.ml.write_subscription_states(self.key, mailinglist_id)
+        self.ml.write_subscription_states(self.key, (mailinglist_id,))
 
         expectation = {
             1: SS.implicit,
