@@ -1,4 +1,5 @@
 #!/usr/bin/env python3
+# pylint: disable=missing-module-docstring
 
 import re
 import urllib.parse
@@ -146,16 +147,15 @@ class TestCoreFrontend(FrontendTest):
               "vera", "werner", "annika", "farin", "akira")
     def test_showuser(self) -> None:
         self.traverse({'description': self.user['display_name']})
-        self.assertTitle("{} {}".format(self.user['given_names'],
-                                        self.user['family_name']))
-        self.assertPresence(self.user['given_names'], div='title')
+        self.assertTitle(self.user['default_name_format'])
+        self.assertPresence(self.user['family_name'], div='title')
 
     @as_users("annika", "paul", "quintus")
     def test_showuser_events(self) -> None:
         if self.user_in("annika"):
             # event admins navigate via event page
             self.traverse("Veranstaltungen", "Große Testakademie",
-                          "Garcia G. Generalis")
+                          "Garcia Generalis")
         elif self.user_in("paul"):
             # core admin
             self.admin_view_profile("garcia")
@@ -164,7 +164,7 @@ class TestCoreFrontend(FrontendTest):
             self.realm_admin_view_profile("garcia", "cde")
 
         self.traverse("Veranstaltungs-Daten")
-        self.assertTitle("Garcia G. Generalis – Veranstaltungs-Daten")
+        self.assertTitle("Garcia Generalis – Veranstaltungs-Daten")
         self.assertPresence("CyberTestAkademie Teilnehmer")
         # part names not shown for one-part events
         self.assertNonPresence("CyberTestAkademie: Teilnehmer")
@@ -216,11 +216,11 @@ class TestCoreFrontend(FrontendTest):
         f['qval_given_names,display_name'] = "Berta"
         self.submit(f)
 
-        self.assertTitle("Bertålotta Beispiel")
+        self.assertTitle(USER_DICT['berta']['default_name_format'])
         self.traverse({'description': 'VCard'})
         vcard = ["BEGIN:VCARD",
                  "VERSION:3.0",
-                 "ADR:;;Im Garten 77;Utopia;;34576;",
+                 "ADR:;;Im Garten 77;Utopia;;34576;Deutschland",
                  "BDAY:1981-02-11",
                  "EMAIL:berta@example.cde",
                  "FN:Bertålotta Beispiel",
@@ -235,7 +235,6 @@ class TestCoreFrontend(FrontendTest):
     @as_users("vera")
     def test_vcard_cde_admin(self) -> None:
         self.admin_view_profile('charly')
-        self.assertTitle("Charly C. Clown")
         self.traverse({'description': 'VCard'})
 
     @as_users("vera")
@@ -295,7 +294,7 @@ class TestCoreFrontend(FrontendTest):
     @as_users("vera")
     def test_adminshowuser(self) -> None:
         self.admin_view_profile('berta')
-        self.assertTitle("Bertålotta Beispiel")
+        self.assertTitle("Bertå Beispiel")
         self.assertPresence("Bei Überweisungen aus dem Ausland achte bitte",
                             div='copy-paste-template')
 
@@ -328,14 +327,12 @@ class TestCoreFrontend(FrontendTest):
     def test_selectpersona(self) -> None:
         self.get('/core/persona/select?kind=admin_persona&phrase=din')
         expectation = {
-            'personas': [{'display_name': 'Daniel',
-                          'email': 'daniel@example.cde',
+            'personas': [{'email': 'daniel@example.cde',
                           'id': 4,
-                          'name': 'Daniel D. Dino'},
-                         {'display_name': 'Ferdinand',
-                          'email': 'ferdinand@example.cde',
+                          'name': 'Daniel Dino'},
+                         {'email': 'ferdinand@example.cde',
                           'id': 6,
-                          'name': 'Ferdinand F. Findus'}]}
+                          'name': 'Ferdinand Findus'}]}
         self.assertEqual(expectation, self.response.json)
         self.get('/core/persona/select?kind=ml_user&phrase=@exam')
         expectation = (1, 2, 3, 4, 5, 6, 7, 9, 10, 11, 13, 14)
@@ -493,17 +490,17 @@ class TestCoreFrontend(FrontendTest):
 
     @as_users("vera")
     def test_adminshowuser_advanced(self) -> None:
-        for phrase, title in (("DB-2-7", "Bertålotta Beispiel"),
-                              ("2", "Bertålotta Beispiel"),
-                              ("Bertålotta Beispiel", "Bertålotta Beispiel"),
-                              ("berta@example.cde", "Bertålotta Beispiel"),
-                              ("anton@example.cde", "Anton Armin A. Administrator"),
-                              ("Spielmanns", "Bertålotta Beispiel"),):
+        for phrase, user in (("DB-2-7", USER_DICT['berta']),
+                              ("2", USER_DICT['berta']),
+                              ("Bertålotta Beispiel", USER_DICT['berta']),
+                              ("berta@example.cde", USER_DICT['berta']),
+                              ("anton@example.cde", USER_DICT['anton']),
+                              ("Spielmanns", USER_DICT['berta']),):
             self.traverse({'href': '^/$'})
             f = self.response.forms['adminshowuserform']
             f['phrase'] = phrase
             self.submit(f)
-            self.assertTitle(title)
+            self.assertTitle(user['default_name_format'])
         self.traverse({'href': '^/$'})
         f = self.response.forms['adminshowuserform']
         f['phrase'] = "nonsense asorecuhasoecurhkgdgdckgdoao"
@@ -535,10 +532,21 @@ class TestCoreFrontend(FrontendTest):
         self.assertPresence("(Zelda)", div='personal-information')
 
     @as_users("vera")
+    def test_automatic_country(self) -> None:
+        self.admin_view_profile('annika')
+        self.traverse({'description': 'Bearbeiten'})
+        f = self.response.forms['changedataform']
+        f['location2'] = "Kabul"
+        self.submit(f)
+        self.assertTitle("Annika Akademieteam")
+        self.assertNonPresence("Afghanistan")
+        self.assertPresence("Deutschland", div='address2')
+
+    @as_users("vera")
     def test_adminchangedata_other(self) -> None:
         self.admin_view_profile('berta')
         self.traverse({'description': 'Bearbeiten'})
-        self.assertTitle("Bertålotta Beispiel bearbeiten")
+        self.assertTitle("Bertå Beispiel bearbeiten")
         f = self.response.forms['changedataform']
         f['display_name'] = "Zelda"
         f['birthday'] = "3.4.1933"
@@ -801,11 +809,11 @@ class TestCoreFrontend(FrontendTest):
     @as_users("vera", "ferdinand")
     def test_cde_admin_reset_password(self) -> None:
         self.realm_admin_view_profile('berta', 'cde')
-        self.assertTitle("Bertålotta Beispiel")
+        self.assertTitle("Bertå Beispiel")
         f = self.response.forms['sendpasswordresetform']
         self.submit(f)
         self.assertPresence("E-Mail abgeschickt.", div='notifications')
-        self.assertTitle("Bertålotta Beispiel")
+        self.assertTitle("Bertå Beispiel")
 
     @as_users("ferdinand", "nina")
     def test_ml_admin_reset_password(self) -> None:
@@ -932,8 +940,8 @@ class TestCoreFrontend(FrontendTest):
             new_privileges, old_privileges, new_password=new_password)
         # Check success.
         self.get('/core/persona/{}/privileges'.format(new_admin["id"]))
-        self.assertTitle("Privilegien ändern für {} {}".format(
-            new_admin["given_names"], new_admin["family_name"]))
+        self.assertTitle("Privilegien ändern für {}".format(
+            new_admin['default_name_format']))
         f = self.response.forms['privilegechangeform']
         old_privileges.update(new_privileges)
         for k, v in old_privileges.items():
@@ -951,8 +959,8 @@ class TestCoreFrontend(FrontendTest):
     def test_change_privileges_dependency_error(self) -> None:
         new_admin = USER_DICT["berta"]
         self.get('/core/persona/{}/privileges'.format(new_admin["id"]))
-        self.assertTitle("Privilegien ändern für {} {}".format(
-            new_admin["given_names"], new_admin["family_name"]))
+        self.assertTitle("Privilegien ändern für {}".format(
+            new_admin["default_name_format"]))
         f = self.response.forms['privilegechangeform']
         f['is_finance_admin'] = True
         f['notes'] = "Berta ist jetzt Praktikant der Finanz Vorstände."
@@ -984,8 +992,8 @@ class TestCoreFrontend(FrontendTest):
             new_privileges, old_privileges)
         # Check success.
         self.get('/core/persona/{}/privileges'.format(new_admin["id"]))
-        self.assertTitle("Privilegien ändern für {} {}".format(
-            new_admin["given_names"], new_admin["family_name"]))
+        self.assertTitle("Privilegien ändern für {}".format(
+            new_admin["default_name_format"]))
         f = self.response.forms['privilegechangeform']
         # Check that old privileges are still active.
         for k, v in old_privileges.items():
@@ -999,8 +1007,8 @@ class TestCoreFrontend(FrontendTest):
         self.submit(f)
         self.traverse(
             {'href': '/core/persona/{}/privileges'.format(new_admin["id"])})
-        self.assertTitle("Privilegien ändern für {} {}".format(
-            new_admin["given_names"], new_admin["family_name"]))
+        self.assertTitle("Privilegien ändern für {}".format(
+            new_admin["default_name_format"]))
         f = self.response.forms['privilegechangeform']
         self.assertNotIn('is_meta_admin', f.fields)
         self.assertNotIn('is_core_admin', f.fields)
@@ -1047,7 +1055,7 @@ class TestCoreFrontend(FrontendTest):
         self._initialize_privilege_change(user, user, user, new_privileges)
         self.login(user)
         self.traverse({'description': "Admin-Änderungen"},
-                      {'description': user["given_names"]})
+                      {'description': "A. Administrator"})
         self.assertPresence(
             "Diese Änderung der Admin-Privilegien wurde von Dir angestoßen",
             div="notifications")
@@ -1077,8 +1085,8 @@ class TestCoreFrontend(FrontendTest):
         self.submit(f)
         self.traverse(
             {'href': '/core/persona/{}/privileges'.format(new_admin["id"])})
-        self.assertTitle("Privilegien ändern für {} {}".format(
-            new_admin["given_names"], new_admin["family_name"]))
+        self.assertTitle("Privilegien ändern für {}".format(
+            new_admin["default_name_format"]))
         f = self.response.forms['privilegechangeform']
         if old_privileges:
             for k, v in old_privileges.items():
@@ -1101,8 +1109,7 @@ class TestCoreFrontend(FrontendTest):
         # Confirm privilege change.
         self.login(admin2)
         self.traverse({'description': "Admin-Änderungen"},
-                      {'description': new_admin["given_names"] + " " +
-                                      new_admin["family_name"]})
+                      {'description': new_admin['family_name']})
         f = self.response.forms["ackprivilegechangeform"]
         self.submit(f)
         self.assertPresence("Änderung wurde übernommen.", div="notifications")
@@ -1129,8 +1136,7 @@ class TestCoreFrontend(FrontendTest):
         # Confirm privilege change.
         self.login(admin2)
         self.traverse({'description': "Admin-Änderungen"},
-                      {'description': new_admin["given_names"] + " " +
-                                      new_admin["family_name"]})
+                      {'description': new_admin['family_name']})
         f = self.response.forms["nackprivilegechangeform"]
         self.submit(f)
         self.assertPresence("Änderung abgelehnt.", div="notifications")
@@ -1293,7 +1299,7 @@ class TestCoreFrontend(FrontendTest):
             self.admin_view_profile('charly')
         else:
             self.realm_admin_view_profile('charly', realm='cde')
-        self.assertTitle("Charly C. Clown")
+        self.assertTitle("Charly Clown")
         self.assertNonPresence("Der Benutzer ist archiviert.")
         self.assertPresence("Zirkusstadt", div='address')
         f = self.response.forms['archivepersonaform']
@@ -1302,21 +1308,21 @@ class TestCoreFrontend(FrontendTest):
         self.assertValidationError(
             "note", "Darf nicht leer sein",
             notification="Archivierungsnotiz muss angegeben werden.")
-        self.assertTitle("Charly C. Clown")
+        self.assertTitle("Charly Clown")
         self.assertNonPresence("Der Benutzer ist archiviert.")
         self.assertPresence("Zirkusstadt", div='address')
         f = self.response.forms['archivepersonaform']
         f['ack_delete'].checked = True
         f['note'] = "Archived for testing."
         self.submit(f)
-        self.assertTitle("Charly C. Clown")
+        self.assertTitle("Charly Clown")
         self.assertPresence("Der Benutzer ist archiviert.", div='archived')
         self.assertNonPresence("Zirkusstadt")
         self.traverse({'description': "Account wiederherstellen"})
         f = self.response.forms['dearchivepersonaform']
         f['new_username'] = "charly@example.cde"
         self.submit(f)
-        self.assertTitle("Charly C. Clown")
+        self.assertTitle("Charly Clown")
         self.assertNonPresence("Der Benutzer ist archiviert.")
 
     @as_users("vera")
@@ -1414,11 +1420,11 @@ class TestCoreFrontend(FrontendTest):
         self.login(user)
         self.traverse({'description': 'Änderungen prüfen'})
         self.assertTitle("Zu prüfende Profiländerungen [1]")
-        self.traverse({'description': 'Bertålotta Ganondorf'},
+        self.traverse({'description': 'Ganondorf'},
                       {'description': 'Änderungen bearbeiten'})
-        self.assertTitle("Bertålotta Ganondorf bearbeiten")
+        self.assertTitle("Bertå Ganondorf bearbeiten")
         self.traverse({'description': 'Änderungen prüfen'},
-                      {'description': 'Bertålotta Ganondorf'})
+                      {'description': 'Ganondorf'})
         f = self.response.forms['ackchangeform']
         self.submit(f)
         self.assertTitle("Zu prüfende Profiländerungen [0]")
@@ -1427,7 +1433,7 @@ class TestCoreFrontend(FrontendTest):
         f['reviewed_by'] = 'DB-22-1'
         self.submit(f)
         self.assertTitle('Nutzerdaten-Log [1–1 von 1]')
-        self.assertPresence("Bertålotta Ganondorf")
+        self.assertPresence("Bertå Ganondorf")
         self.logout()
         user = USER_DICT["berta"]
         self.login(user)
@@ -1440,7 +1446,7 @@ class TestCoreFrontend(FrontendTest):
     def test_history(self) -> None:
         self.admin_view_profile('berta')
         self.traverse({'href': '/core/persona/2/adminchange'})
-        self.assertTitle("Bertålotta Beispiel bearbeiten")
+        self.assertTitle("Bertå Beispiel bearbeiten")
         f = self.response.forms['changedataform']
         f['display_name'] = "Zelda"
         f['birthday'] = "3.4.1933"
@@ -1522,12 +1528,17 @@ class TestCoreFrontend(FrontendTest):
         self.submit(f)
         self.assertTitle("Bereichsänderung für Emilia E. Eventis")
         f = self.response.forms['promotionform']
+        self.submit(f, check_notification=False)
+        self.assertValidationError('change_note', "Darf nicht leer sein.")
+        f['change_note'] = change_note = "Hat an einer Akademie teilgenommen."
         self.submit(f)
         self.assertTitle("Emilia E. Eventis")
         self.assertPresence("0,00 €", div='balance')
         self.assertCheckbox(True, "paper_expuls_checkbox")
         self.assertNonPresence("CdE-Mitglied", div="cde-membership")
         self.assertNonPresence("Probemitgliedschaft", div="cde-membership")
+        self.traverse("Änderungshistorie")
+        self.assertPresence(change_note, div="generation2")
 
         # Do another promotion, this time granting trial membership.
         self.admin_view_profile('nina')
@@ -1540,33 +1551,46 @@ class TestCoreFrontend(FrontendTest):
         self.assertTitle("Bereichsänderung für Nina Neubauer")
         f = self.response.forms['promotionform']
         f['trial_member'].checked = True
+        f['change_note'] = "Per Vorstandsbeschluss aufgenommen."
         self.submit(f)
         self.assertTitle("Nina Neubauer")
         self.assertPresence("0,00 €", div='balance')
         self.assertPresence("CdE-Mitglied", div="cde-membership")
         self.assertPresence("Probemitgliedschaft", div="cde-membership")
 
+        # check for correct welcome mail
+        mail = self.fetch_mail_content()
+        self.assertIn(USER_DICT['nina']['display_name'], mail)
+        self.assertIn("Ein herzliches Willkommen", mail)
+        self.assertIn("zum ersten Mal in unserer Datenbank anmeldest", mail)
+        self.assertIn("kostenlos", mail)  # check trial membership
+
     @as_users("vera")
     def test_nontrivial_promotion(self) -> None:
         self.admin_view_profile('kalif')
         self.traverse({'description': 'Bereich hinzufügen'})
-        self.assertTitle("Bereichsänderung für Kalif ibn al-Ḥasan Karabatschi")
+        self.assertTitle("Bereichsänderung für Kalif Karabatschi")
         f = self.response.forms['realmselectionform']
         f['target_realm'] = "event"
         self.submit(f)
-        self.assertTitle("Bereichsänderung für Kalif ibn al-Ḥasan Karabatschi")
+        self.assertTitle("Bereichsänderung für Kalif Karabatschi")
         f = self.response.forms['promotionform']
         # First check error handling by entering an invalid birthday
         f['birthday'] = "foobar"
         self.submit(f, check_notification=False)
         self.assertValidationError("birthday", "Ungültige Eingabe für ein Datum")
-        self.assertTitle("Bereichsänderung für Kalif ibn al-Ḥasan Karabatschi")
+        self.assertValidationError('change_note', "Darf nicht leer sein.")
+        self.assertTitle("Bereichsänderung für Kalif Karabatschi")
         # Now, do it right
         f['birthday'] = "21.6.1977"
         f['gender'] = 1
+        f['change_note'] = "Komplizierte Aufnahme"
         self.submit(f)
-        self.assertTitle("Kalif ibn al-Ḥasan Karabatschi")
+        self.assertTitle("Kalif Karabatschi")
         self.assertPresence("21.06.1977", div='personal-information')
+        # check that no welcome mail is sent - this is for cde promotion only
+        with self.assertRaises(IndexError):
+            self.fetch_mail_content()
 
     @as_users("vera")
     def test_ignore_warnings_postal_code(self) -> None:
@@ -1659,8 +1683,7 @@ class TestCoreFrontend(FrontendTest):
     def test_genesis_event(self) -> None:
         self._genesis_request(self.EVENT_GENESIS_DATA)
 
-        user = USER_DICT['vera']
-        self.login(user)
+        self.login('vera')
         self.traverse({'description': 'Accountanfrage'})
         self.assertTitle("Accountanfragen")
         self.assertPresence("zelda@example.cde", div='request-1001')
@@ -1766,6 +1789,9 @@ class TestCoreFrontend(FrontendTest):
         f = self.response.forms['genesisform']
         for field, entry in self.CDE_GENESIS_DATA.items():
             f[field] = entry
+        self.submit(f, check_notification=False)
+        self.assertValidationError('attachment')
+        f = self.response.forms['genesisform']
         f['birth_name'] = "Ganondorf"
         f['notes'] = ""  # Do not send this to test upload permanance.
         with open(self.testfile_dir / "form.pdf", 'rb') as datafile:
@@ -1777,6 +1803,10 @@ class TestCoreFrontend(FrontendTest):
         self.assertPresence("Anhang my_participation_certificate.pdf")
         f = self.response.forms['genesisform']
         f['notes'] = "Gimme!"
+        f['birthday'] = ""
+        self.submit(f, check_notification=False)
+        self.assertValidationError("birthday", "Darf nicht leer sein.")
+        f['birthday'] = self.CDE_GENESIS_DATA['birthday']
         self.submit(f)
         link = self.fetch_link()
         self.get(link)
@@ -1796,13 +1826,24 @@ class TestCoreFrontend(FrontendTest):
         self.traverse({'href': '/core/genesis/1001/show'})
         self.assertPresence("Ganondorf")
         self.assertNonPresence("Zickzack")
+        self.assertNonPresence("PfingstAkademie")
         self.traverse({'href': '/core/genesis/1001/modify'})
         self.assertTitle("Accountanfrage bearbeiten")
         f = self.response.forms['genesismodifyform']
         f['birth_name'] = "Zickzack"
+        f['pevent_id'] = 1
         self.submit(f)
         self.assertPresence("Zickzack")
         self.assertNonPresence("Ganondorf")
+        self.assertPresence("PfingstAkademie 2014")
+
+        self.traverse({'href': '/core/genesis/1001/modify'})
+        self.assertTitle("Accountanfrage bearbeiten")
+        f = self.response.forms['genesismodifyform']
+        f['pcourse_id'] = 2
+        self.submit(f)
+        self.assertPresence("Goethe")
+
         self.traverse({'href': '/core/genesis/1001/modify'})
         self.assertTitle("Accountanfrage bearbeiten")
         f = self.response.forms['genesismodifyform']
@@ -1873,6 +1914,8 @@ class TestCoreFrontend(FrontendTest):
         self.assertCheckbox(True, "paper_expuls_checkbox")
         self.assertPresence("12345")
         self.assertPresence("Zickzack")
+        self.assertPresence("PfingstAkademie 2014")
+        self.assertPresence("Goethe")
         self.traverse({'href': '/cde'})
         self.assertTitle('CdE-Mitgliederbereich')
         self.traverse({'description': 'Sonstiges'})
@@ -2064,6 +2107,7 @@ class TestCoreFrontend(FrontendTest):
         f['target_realm'] = "assembly"
         self.submit(f)
         f = self.response.forms['promotionform']
+        f['change_note'] = promotion_change_note = "trivial promotion"
         self.submit(f)
         logs.append((1011, const.CoreLogCodes.realm_change))
 
@@ -2084,6 +2128,6 @@ class TestCoreFrontend(FrontendTest):
                       const.CoreLogCodes.realm_change.value,
                       const.CoreLogCodes.username_change.value]
         self.submit(f)
-        self.assertPresence("Bereiche geändert.")
+        self.assertPresence(promotion_change_note)
         self.assertPresence("zelda@example.cde")
         self.assertPresence("bertalotta@example.cde")
