@@ -1,4 +1,5 @@
 #!/usr/bin/env python3
+# pylint: disable=missing-module-docstring
 
 import collections.abc
 import datetime
@@ -528,7 +529,7 @@ class TestCron(CronTest):
         self.assertEqual(
             mm_lists['witz'].subscribe.call_args_list,
             [umcall('new-anton@example.cde',
-                    display_name='Anton Armin A. Administrator',
+                    display_name='Anton Administrator',
                     pre_approved=True, pre_confirmed=True, pre_verified=True)])
         self.assertEqual(
             mm_lists['witz'].unsubscribe.call_args_list,
