@@ -1,4 +1,6 @@
 #!/usr/bin/env python3
+# pylint: disable=missing-module-docstring
+
 import os
 import pathlib
 import shutil
@@ -69,7 +71,7 @@ class TestOffline(FrontendTest):
                           {'href': 'event/event/1/registration/1/show'},
                           {'href': 'core/persona/1/show'},
                           {'href': 'core/persona/1/adminchange'})
-            self.assertTitle('Anton Armin A. Administrator bearbeiten')
+            self.assertTitle('Anton Administrator bearbeiten')
             f = self.response.forms['changedataform']
             f['display_name'] = "Zelda"
             f['birthday'] = "3.4.1933"
