@@ -1,4 +1,6 @@
 #!/usr/bin/env python3
+# pylint: disable=missing-module-docstring
+
 import datetime
 import secrets
 from typing import List, NamedTuple, Sequence, Optional, cast
@@ -183,7 +185,7 @@ class TestMultiSessionFrontend(MultiAppFrontendTest):
             keys.append(self.app.cookies[session_cookie])
             # Check that we are correctly logged in.
             self.get("/core/self/show")
-            self.assertTitle(f"{user['given_names']} {user['family_name']}")
+            self.assertTitle(user['default_name_format'])
             self.assertNotIn('loginform', self.response.forms)
         self.assertEqual(len(set(keys)), len(keys))
 
@@ -205,7 +207,7 @@ class TestMultiSessionFrontend(MultiAppFrontendTest):
             self.switch_app(i)
             with self.subTest(app_index=i):
                 self.get("/core/self/show")
-                self.assertTitle(f"{user['given_names']} {user['family_name']}")
+                self.assertTitle(user['default_name_format'])
                 self.assertNotIn('loginform', self.response.forms)
 
         # Now terminate all sessions and check that they are all inactive.
